@@ -1,25 +1,18 @@
 module github.com/s7techlab/cckit
 
-go 1.12
+go 1.13
 
 require (
-	github.com/Knetic/govaluate v3.0.0+incompatible // indirect
-	github.com/fsouza/go-dockerclient v1.4.0 // indirect
-	github.com/gogo/protobuf v1.2.1
-	github.com/golang/mock v1.2.0 // indirect
-	github.com/golang/protobuf v1.3.1
-	github.com/grpc-ecosystem/grpc-gateway v1.9.0
-	github.com/hyperledger/fabric v1.4.0
-	github.com/hyperledger/fabric-amcl v0.0.0-20181230093703-5ccba6eab8d6 // indirect
-	github.com/miekg/pkcs11 v1.0.2 // indirect
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/golang/protobuf v1.3.2
+	github.com/grpc-ecosystem/grpc-gateway v1.11.1
+	github.com/hyperledger/fabric v1.4.4
 	github.com/mwitkow/go-proto-validators v0.0.0-20190212092829-1f388280e944
 	github.com/onsi/ginkgo v1.8.0
 	github.com/onsi/gomega v1.5.0
 	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7
 	github.com/pkg/errors v0.8.1
-	github.com/s7techlab/hlf-sdk-go v0.1.3
-	github.com/spf13/viper v1.4.0 // indirect
-	golang.org/x/net v0.0.0-20190522155817-f3200d17e092
-	google.golang.org/genproto v0.0.0-20190307195333-5fe7a883aa19
-	google.golang.org/grpc v1.21.0
+	github.com/s7techlab/hlf-sdk-go v0.5.0
+	google.golang.org/genproto v0.0.0-20190716160619-c506a9f90610
+	google.golang.org/grpc v1.23.0
 )
