@@ -1,26 +1,19 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // source: service.proto
 
-/*
-Package service is a generated protocol buffer package.
+// Commercial paper service
 
-It is generated from these files:
-	service.proto
-
-It has these top-level messages:
-*/
 package service
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
-import _ "google.golang.org/genproto/googleapis/api/annotations"
-import google_protobuf1 "github.com/golang/protobuf/ptypes/empty"
-import schema "github.com/s7techlab/cckit/examples/cpaper_asservice/schema"
-
 import (
-	context "golang.org/x/net/context"
+	context "context"
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	empty "github.com/golang/protobuf/ptypes/empty"
+	schema "github.com/s7techlab/cckit/examples/cpaper_asservice/schema"
+	_ "google.golang.org/genproto/googleapis/api/annotations"
 	grpc "google.golang.org/grpc"
+	math "math"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -32,7 +25,38 @@ var _ = math.Inf
 // is compatible with the proto package it is being compiled against.
 // A compilation error at this line likely means your copy of the
 // proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
+
+func init() { proto.RegisterFile("service.proto", fileDescriptor_a0b84a42fa06f626) }
+
+var fileDescriptor_a0b84a42fa06f626 = []byte{
+	// 386 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x92, 0x41, 0xcb, 0xd3, 0x30,
+	0x18, 0xc7, 0xd1, 0xd7, 0xb7, 0xc3, 0xc8, 0x9c, 0x3c, 0xca, 0x74, 0xdd, 0x44, 0xec, 0x45, 0xd8,
+	0xa1, 0x01, 0x3d, 0x08, 0x1e, 0x37, 0xc7, 0x1c, 0xec, 0x20, 0x3b, 0x89, 0x1e, 0x24, 0x4d, 0x1f,
+	0xb7, 0x60, 0xd3, 0x94, 0x24, 0x95, 0x95, 0xb1, 0x8b, 0x27, 0xef, 0x7e, 0x34, 0xbf, 0x82, 0x1f,
+	0x44, 0x9a, 0x26, 0x08, 0x42, 0x71, 0xf0, 0x9e, 0xca, 0xbf, 0x4f, 0xfe, 0xbf, 0x1f, 0xe1, 0x09,
+	0x19, 0x1a, 0xd4, 0xdf, 0x04, 0xc7, 0xb4, 0xd2, 0xca, 0x2a, 0x18, 0xf8, 0x18, 0xcf, 0xf6, 0x4a,
+	0xed, 0x0b, 0xa4, 0xac, 0x12, 0x94, 0x95, 0xa5, 0xb2, 0xcc, 0x0a, 0x55, 0x9a, 0xee, 0x58, 0x3c,
+	0xf5, 0x53, 0x97, 0xb2, 0xfa, 0x0b, 0x45, 0x59, 0xd9, 0xc6, 0x0f, 0xdf, 0xed, 0x85, 0x3d, 0xd4,
+	0x59, 0xca, 0x95, 0xa4, 0xe6, 0xb5, 0x45, 0x7e, 0x28, 0x58, 0x46, 0x39, 0xff, 0x2a, 0x2c, 0xc5,
+	0x23, 0x93, 0x55, 0x81, 0x86, 0xf2, 0x8a, 0x55, 0xa8, 0x3f, 0x33, 0xe3, 0x85, 0xd4, 0xf0, 0x03,
+	0x4a, 0xe6, 0x3f, 0x1d, 0xe9, 0xe5, 0x8f, 0x6b, 0x12, 0x2d, 0xdf, 0xb7, 0x07, 0x61, 0x4b, 0xee,
+	0x6c, 0x85, 0xb1, 0x30, 0x4e, 0x3b, 0x75, 0x1a, 0xd4, 0xe9, 0xaa, 0x55, 0xc7, 0xd3, 0xd4, 0x37,
+	0x97, 0x4a, 0x4a, 0xd4, 0x5c, 0xb0, 0xc2, 0x15, 0xdb, 0x52, 0x32, 0xfa, 0xfe, 0xeb, 0xf7, 0xcf,
+	0xdb, 0x77, 0x61, 0xe0, 0xb5, 0x90, 0x91, 0xab, 0x35, 0x5a, 0x98, 0xf4, 0x94, 0x36, 0x79, 0xfc,
+	0xb8, 0x67, 0x94, 0xbc, 0x70, 0xac, 0xe7, 0xf0, 0xcc, 0xb3, 0xe8, 0x49, 0x18, 0x53, 0xa3, 0x3e,
+	0xd3, 0x53, 0x77, 0xa5, 0xb2, 0x96, 0x19, 0xea, 0x33, 0x7c, 0x24, 0xa3, 0x35, 0xda, 0x45, 0xb3,
+	0x3a, 0x5a, 0xd4, 0x25, 0x2b, 0x36, 0x39, 0x40, 0x80, 0xfe, 0xfd, 0xd7, 0x2f, 0x8a, 0x9d, 0xe8,
+	0x11, 0x40, 0x10, 0xe1, 0xd1, 0x8a, 0x9c, 0x9e, 0x44, 0x7e, 0x86, 0x0f, 0xe4, 0x7a, 0xd3, 0x5a,
+	0x61, 0x16, 0xda, 0x2e, 0xfe, 0x83, 0xe8, 0x67, 0x3f, 0x71, 0x6c, 0x48, 0x86, 0x81, 0xed, 0xee,
+	0xf0, 0xe6, 0xd6, 0x1c, 0x76, 0xe4, 0x6a, 0x51, 0x37, 0x10, 0x87, 0xe6, 0xa2, 0x6e, 0x2e, 0xa6,
+	0x8e, 0x1d, 0xf5, 0x41, 0x72, 0x2f, 0x50, 0xb3, 0xba, 0x69, 0x99, 0x9f, 0x48, 0xb4, 0xc3, 0x1c,
+	0x51, 0xc2, 0xd3, 0x50, 0xed, 0xf2, 0xc5, 0xe4, 0x89, 0x23, 0x3f, 0x4c, 0xee, 0x07, 0xb2, 0x76,
+	0xfd, 0x16, 0x8e, 0x24, 0x7a, 0x8b, 0x05, 0x5a, 0xbc, 0xc9, 0x36, 0xe7, 0xff, 0xdb, 0x66, 0x16,
+	0xb9, 0xf7, 0xf6, 0xea, 0x4f, 0x00, 0x00, 0x00, 0xff, 0xff, 0x32, 0x1f, 0x1d, 0x98, 0x30, 0x03,
+	0x00, 0x00,
+}
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ context.Context
@@ -42,15 +66,23 @@ var _ grpc.ClientConn
 // is compatible with the grpc package it is being compiled against.
 const _ = grpc.SupportPackageIsVersion4
 
-// Client API for CPaper service
-
+// CPaperClient is the client API for CPaper service.
+//
+// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
 type CPaperClient interface {
-	List(ctx context.Context, in *google_protobuf1.Empty, opts ...grpc.CallOption) (*schema.CommercialPaperList, error)
+	// List method returns all registered commercial papers
+	List(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*schema.CommercialPaperList, error)
+	// Get method returns commercial paper data by id
 	Get(ctx context.Context, in *schema.CommercialPaperId, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
+	// GetByExternalId
 	GetByExternalId(ctx context.Context, in *schema.ExternalId, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
+	// Issue commercial paper
 	Issue(ctx context.Context, in *schema.IssueCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
+	// Buy commercial paper
 	Buy(ctx context.Context, in *schema.BuyCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
+	// Redeem commercial paper
 	Redeem(ctx context.Context, in *schema.RedeemCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
+	// Delete commercial paper
 	Delete(ctx context.Context, in *schema.CommercialPaperId, opts ...grpc.CallOption) (*schema.CommercialPaper, error)
 }
 
@@ -62,9 +94,9 @@ func NewCPaperClient(cc *grpc.ClientConn) CPaperClient {
 	return &cPaperClient{cc}
 }
 
-func (c *cPaperClient) List(ctx context.Context, in *google_protobuf1.Empty, opts ...grpc.CallOption) (*schema.CommercialPaperList, error) {
+func (c *cPaperClient) List(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*schema.CommercialPaperList, error) {
 	out := new(schema.CommercialPaperList)
-	err := grpc.Invoke(ctx, "/service.CPaper/List", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/List", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -73,7 +105,7 @@ func (c *cPaperClient) List(ctx context.Context, in *google_protobuf1.Empty, opt
 
 func (c *cPaperClient) Get(ctx context.Context, in *schema.CommercialPaperId, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/Get", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/Get", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -82,7 +114,7 @@ func (c *cPaperClient) Get(ctx context.Context, in *schema.CommercialPaperId, op
 
 func (c *cPaperClient) GetByExternalId(ctx context.Context, in *schema.ExternalId, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/GetByExternalId", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/GetByExternalId", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -91,7 +123,7 @@ func (c *cPaperClient) GetByExternalId(ctx context.Context, in *schema.ExternalI
 
 func (c *cPaperClient) Issue(ctx context.Context, in *schema.IssueCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/Issue", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/Issue", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -100,7 +132,7 @@ func (c *cPaperClient) Issue(ctx context.Context, in *schema.IssueCommercialPape
 
 func (c *cPaperClient) Buy(ctx context.Context, in *schema.BuyCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/Buy", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/Buy", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -109,7 +141,7 @@ func (c *cPaperClient) Buy(ctx context.Context, in *schema.BuyCommercialPaper, o
 
 func (c *cPaperClient) Redeem(ctx context.Context, in *schema.RedeemCommercialPaper, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/Redeem", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/Redeem", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -118,22 +150,28 @@ func (c *cPaperClient) Redeem(ctx context.Context, in *schema.RedeemCommercialPa
 
 func (c *cPaperClient) Delete(ctx context.Context, in *schema.CommercialPaperId, opts ...grpc.CallOption) (*schema.CommercialPaper, error) {
 	out := new(schema.CommercialPaper)
-	err := grpc.Invoke(ctx, "/service.CPaper/Delete", in, out, c.cc, opts...)
+	err := c.cc.Invoke(ctx, "/service.CPaper/Delete", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-// Server API for CPaper service
-
+// CPaperServer is the server API for CPaper service.
 type CPaperServer interface {
-	List(context.Context, *google_protobuf1.Empty) (*schema.CommercialPaperList, error)
+	// List method returns all registered commercial papers
+	List(context.Context, *empty.Empty) (*schema.CommercialPaperList, error)
+	// Get method returns commercial paper data by id
 	Get(context.Context, *schema.CommercialPaperId) (*schema.CommercialPaper, error)
+	// GetByExternalId
 	GetByExternalId(context.Context, *schema.ExternalId) (*schema.CommercialPaper, error)
+	// Issue commercial paper
 	Issue(context.Context, *schema.IssueCommercialPaper) (*schema.CommercialPaper, error)
+	// Buy commercial paper
 	Buy(context.Context, *schema.BuyCommercialPaper) (*schema.CommercialPaper, error)
+	// Redeem commercial paper
 	Redeem(context.Context, *schema.RedeemCommercialPaper) (*schema.CommercialPaper, error)
+	// Delete commercial paper
 	Delete(context.Context, *schema.CommercialPaperId) (*schema.CommercialPaper, error)
 }
 
@@ -142,7 +180,7 @@ func RegisterCPaperServer(s *grpc.Server, srv CPaperServer) {
 }
 
 func _CPaper_List_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(google_protobuf1.Empty)
+	in := new(empty.Empty)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
@@ -154,7 +192,7 @@ func _CPaper_List_Handler(srv interface{}, ctx context.Context, dec func(interfa
 		FullMethod: "/service.CPaper/List",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(CPaperServer).List(ctx, req.(*google_protobuf1.Empty))
+		return srv.(CPaperServer).List(ctx, req.(*empty.Empty))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -302,34 +340,4 @@ var _CPaper_serviceDesc = grpc.ServiceDesc{
 	},
 	Streams:  []grpc.StreamDesc{},
 	Metadata: "service.proto",
-}
-
-func init() { proto.RegisterFile("service.proto", fileDescriptor0) }
-
-var fileDescriptor0 = []byte{
-	// 380 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x92, 0xd1, 0xaa, 0xd3, 0x30,
-	0x18, 0xc7, 0xc1, 0xcd, 0x0e, 0x23, 0x73, 0xf2, 0xa9, 0x13, 0xbb, 0x89, 0xd8, 0x1b, 0xc1, 0x8b,
-	0x06, 0xf4, 0xc2, 0xfb, 0xcd, 0x31, 0x07, 0x03, 0x45, 0x10, 0xc4, 0x1b, 0x49, 0xd3, 0xcf, 0x2d,
-	0xd8, 0x34, 0x25, 0x49, 0x65, 0x65, 0xec, 0xc6, 0x57, 0xf0, 0xd1, 0x7c, 0x85, 0xf3, 0x1e, 0xe7,
-	0xd0, 0x34, 0xe1, 0xc0, 0x81, 0x72, 0x06, 0xe7, 0xaa, 0xfc, 0xfb, 0xe5, 0xff, 0xfb, 0x11, 0xbe,
-	0x90, 0xb1, 0x41, 0xfd, 0x47, 0x70, 0x4c, 0x2b, 0xad, 0xac, 0x82, 0x91, 0x8f, 0xf1, 0x7c, 0xa7,
-	0xd4, 0xae, 0x40, 0xca, 0x2a, 0x41, 0x59, 0x59, 0x2a, 0xcb, 0xac, 0x50, 0xa5, 0xe9, 0x8e, 0xc5,
-	0x33, 0x3f, 0x75, 0x29, 0xab, 0x7f, 0x51, 0x94, 0x95, 0x6d, 0xfc, 0xf0, 0xd3, 0x4e, 0xd8, 0x7d,
-	0x9d, 0xa5, 0x5c, 0x49, 0x6a, 0x3e, 0x58, 0xe4, 0xfb, 0x82, 0x65, 0x94, 0xf3, 0xdf, 0xc2, 0x52,
-	0x3c, 0x30, 0x59, 0x15, 0x68, 0x28, 0xaf, 0x58, 0x85, 0xfa, 0x27, 0x33, 0x5e, 0x48, 0x0d, 0xdf,
-	0xa3, 0x64, 0xfe, 0xd3, 0x91, 0xde, 0x5d, 0x0e, 0x49, 0xb4, 0xfc, 0xd2, 0x1e, 0x84, 0x2d, 0x19,
-	0x6e, 0x85, 0xb1, 0x30, 0x4d, 0x3b, 0x75, 0x1a, 0xd4, 0xe9, 0xaa, 0x55, 0xc7, 0xb3, 0xd4, 0x37,
-	0x97, 0x4a, 0x4a, 0xd4, 0x5c, 0xb0, 0xc2, 0x15, 0xdb, 0x52, 0x32, 0xf9, 0xfb, 0xff, 0xe2, 0xdf,
-	0xbd, 0x07, 0x30, 0xf2, 0x5a, 0xc8, 0xc8, 0x60, 0x8d, 0x16, 0x5e, 0xf4, 0x94, 0x36, 0x79, 0xfc,
-	0xbc, 0x67, 0x94, 0xbc, 0x71, 0xac, 0xd7, 0xf0, 0xca, 0xb3, 0xe8, 0x51, 0x18, 0x53, 0xa3, 0x3e,
-	0xd1, 0x63, 0x77, 0xa5, 0xb2, 0x96, 0x19, 0xea, 0x13, 0xfc, 0x20, 0x93, 0x35, 0xda, 0x45, 0xb3,
-	0x3a, 0x58, 0xd4, 0x25, 0x2b, 0x36, 0x39, 0x40, 0x80, 0x5e, 0xff, 0xeb, 0x17, 0xc5, 0x4e, 0xf4,
-	0x14, 0x20, 0x88, 0xf0, 0x60, 0x45, 0x4e, 0x8f, 0x22, 0x3f, 0xc1, 0x37, 0x72, 0x7f, 0xd3, 0x5a,
-	0x61, 0x1e, 0xda, 0x2e, 0xde, 0x40, 0xf4, 0xb3, 0x9f, 0x39, 0xf6, 0x24, 0x19, 0x07, 0xb6, 0xbb,
-	0x03, 0x7c, 0x26, 0x83, 0x45, 0xdd, 0x40, 0x1c, 0x6a, 0x8b, 0xba, 0x39, 0x1b, 0xf9, 0xc4, 0x21,
-	0xc7, 0xc9, 0xc3, 0x80, 0xcc, 0xea, 0x06, 0xbe, 0x93, 0xe8, 0x2b, 0xe6, 0x88, 0x12, 0x5e, 0x86,
-	0x5e, 0x97, 0xcf, 0xc6, 0x4e, 0x1d, 0xf6, 0x71, 0xf2, 0x28, 0x60, 0x75, 0xc7, 0x43, 0x12, 0x7d,
-	0xc4, 0x02, 0x2d, 0xde, 0x65, 0x89, 0x6f, 0x6f, 0x5b, 0x62, 0x16, 0xb9, 0x67, 0xf6, 0xfe, 0x2a,
-	0x00, 0x00, 0xff, 0xff, 0x1d, 0xa1, 0x70, 0x55, 0x27, 0x03, 0x00, 0x00,
 }
