@@ -1,27 +1,15 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // source: chaincode.proto
 
-/*
-Package service is a generated protocol buffer package.
-
-It is generated from these files:
-	chaincode.proto
-
-It has these top-level messages:
-	ChaincodeInput
-	ChaincodeLocator
-*/
 package service
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
-import protos "github.com/hyperledger/fabric/protos/peer"
-import protos1 "github.com/hyperledger/fabric/protos/peer"
-
 import (
-	context "golang.org/x/net/context"
+	context "context"
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	peer "github.com/hyperledger/fabric/protos/peer"
 	grpc "google.golang.org/grpc"
+	math "math"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -33,26 +21,74 @@ var _ = math.Inf
 // is compatible with the proto package it is being compiled against.
 // A compilation error at this line likely means your copy of the
 // proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
+
+type InvocationType int32
+
+const (
+	InvocationType_QUERY  InvocationType = 0
+	InvocationType_INVOKE InvocationType = 1
+)
+
+var InvocationType_name = map[int32]string{
+	0: "QUERY",
+	1: "INVOKE",
+}
+
+var InvocationType_value = map[string]int32{
+	"QUERY":  0,
+	"INVOKE": 1,
+}
+
+func (x InvocationType) String() string {
+	return proto.EnumName(InvocationType_name, int32(x))
+}
+
+func (InvocationType) EnumDescriptor() ([]byte, []int) {
+	return fileDescriptor_97136ef4b384cc22, []int{0}
+}
 
 type ChaincodeInput struct {
 	// Chaincode name
-	Chaincode string `protobuf:"bytes,1,opt,name=chaincode" json:"chaincode,omitempty"`
+	Chaincode string `protobuf:"bytes,1,opt,name=chaincode,proto3" json:"chaincode,omitempty"`
 	// Channel name
-	Channel string `protobuf:"bytes,2,opt,name=channel" json:"channel,omitempty"`
+	Channel string `protobuf:"bytes,2,opt,name=channel,proto3" json:"channel,omitempty"`
 	// Input contains the arguments for invocation.
 	Args [][]byte `protobuf:"bytes,3,rep,name=args,proto3" json:"args,omitempty"`
 	// TransientMap contains data (e.g. cryptographic material) that might be used
 	// to implement some form of application-level confidentiality. The contents
 	// of this field are supposed to always be omitted from the transaction and
 	// excluded from the ledger.
-	Transient map[string][]byte `protobuf:"bytes,4,rep,name=transient" json:"transient,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	Transient            map[string][]byte `protobuf:"bytes,4,rep,name=transient,proto3" json:"transient,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	XXX_NoUnkeyedLiteral struct{}          `json:"-"`
+	XXX_unrecognized     []byte            `json:"-"`
+	XXX_sizecache        int32             `json:"-"`
 }
 
-func (m *ChaincodeInput) Reset()                    { *m = ChaincodeInput{} }
-func (m *ChaincodeInput) String() string            { return proto.CompactTextString(m) }
-func (*ChaincodeInput) ProtoMessage()               {}
-func (*ChaincodeInput) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{0} }
+func (m *ChaincodeInput) Reset()         { *m = ChaincodeInput{} }
+func (m *ChaincodeInput) String() string { return proto.CompactTextString(m) }
+func (*ChaincodeInput) ProtoMessage()    {}
+func (*ChaincodeInput) Descriptor() ([]byte, []int) {
+	return fileDescriptor_97136ef4b384cc22, []int{0}
+}
+
+func (m *ChaincodeInput) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_ChaincodeInput.Unmarshal(m, b)
+}
+func (m *ChaincodeInput) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_ChaincodeInput.Marshal(b, m, deterministic)
+}
+func (m *ChaincodeInput) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ChaincodeInput.Merge(m, src)
+}
+func (m *ChaincodeInput) XXX_Size() int {
+	return xxx_messageInfo_ChaincodeInput.Size(m)
+}
+func (m *ChaincodeInput) XXX_DiscardUnknown() {
+	xxx_messageInfo_ChaincodeInput.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_ChaincodeInput proto.InternalMessageInfo
 
 func (m *ChaincodeInput) GetChaincode() string {
 	if m != nil {
@@ -84,15 +120,38 @@ func (m *ChaincodeInput) GetTransient() map[string][]byte {
 
 type ChaincodeLocator struct {
 	// Chaincode name
-	Chaincode string `protobuf:"bytes,1,opt,name=chaincode" json:"chaincode,omitempty"`
+	Chaincode string `protobuf:"bytes,1,opt,name=chaincode,proto3" json:"chaincode,omitempty"`
 	// Channel name
-	Channel string `protobuf:"bytes,2,opt,name=channel" json:"channel,omitempty"`
+	Channel              string   `protobuf:"bytes,2,opt,name=channel,proto3" json:"channel,omitempty"`
+	XXX_NoUnkeyedLiteral struct{} `json:"-"`
+	XXX_unrecognized     []byte   `json:"-"`
+	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *ChaincodeLocator) Reset()                    { *m = ChaincodeLocator{} }
-func (m *ChaincodeLocator) String() string            { return proto.CompactTextString(m) }
-func (*ChaincodeLocator) ProtoMessage()               {}
-func (*ChaincodeLocator) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{1} }
+func (m *ChaincodeLocator) Reset()         { *m = ChaincodeLocator{} }
+func (m *ChaincodeLocator) String() string { return proto.CompactTextString(m) }
+func (*ChaincodeLocator) ProtoMessage()    {}
+func (*ChaincodeLocator) Descriptor() ([]byte, []int) {
+	return fileDescriptor_97136ef4b384cc22, []int{1}
+}
+
+func (m *ChaincodeLocator) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_ChaincodeLocator.Unmarshal(m, b)
+}
+func (m *ChaincodeLocator) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_ChaincodeLocator.Marshal(b, m, deterministic)
+}
+func (m *ChaincodeLocator) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ChaincodeLocator.Merge(m, src)
+}
+func (m *ChaincodeLocator) XXX_Size() int {
+	return xxx_messageInfo_ChaincodeLocator.Size(m)
+}
+func (m *ChaincodeLocator) XXX_DiscardUnknown() {
+	xxx_messageInfo_ChaincodeLocator.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_ChaincodeLocator proto.InternalMessageInfo
 
 func (m *ChaincodeLocator) GetChaincode() string {
 	if m != nil {
@@ -108,9 +167,92 @@ func (m *ChaincodeLocator) GetChannel() string {
 	return ""
 }
 
+type ChaincodeExec struct {
+	Type                 InvocationType  `protobuf:"varint,1,opt,name=type,proto3,enum=service.InvocationType" json:"type,omitempty"`
+	Input                *ChaincodeInput `protobuf:"bytes,2,opt,name=input,proto3" json:"input,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}        `json:"-"`
+	XXX_unrecognized     []byte          `json:"-"`
+	XXX_sizecache        int32           `json:"-"`
+}
+
+func (m *ChaincodeExec) Reset()         { *m = ChaincodeExec{} }
+func (m *ChaincodeExec) String() string { return proto.CompactTextString(m) }
+func (*ChaincodeExec) ProtoMessage()    {}
+func (*ChaincodeExec) Descriptor() ([]byte, []int) {
+	return fileDescriptor_97136ef4b384cc22, []int{2}
+}
+
+func (m *ChaincodeExec) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_ChaincodeExec.Unmarshal(m, b)
+}
+func (m *ChaincodeExec) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_ChaincodeExec.Marshal(b, m, deterministic)
+}
+func (m *ChaincodeExec) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ChaincodeExec.Merge(m, src)
+}
+func (m *ChaincodeExec) XXX_Size() int {
+	return xxx_messageInfo_ChaincodeExec.Size(m)
+}
+func (m *ChaincodeExec) XXX_DiscardUnknown() {
+	xxx_messageInfo_ChaincodeExec.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_ChaincodeExec proto.InternalMessageInfo
+
+func (m *ChaincodeExec) GetType() InvocationType {
+	if m != nil {
+		return m.Type
+	}
+	return InvocationType_QUERY
+}
+
+func (m *ChaincodeExec) GetInput() *ChaincodeInput {
+	if m != nil {
+		return m.Input
+	}
+	return nil
+}
+
 func init() {
+	proto.RegisterEnum("service.InvocationType", InvocationType_name, InvocationType_value)
 	proto.RegisterType((*ChaincodeInput)(nil), "service.ChaincodeInput")
+	proto.RegisterMapType((map[string][]byte)(nil), "service.ChaincodeInput.TransientEntry")
 	proto.RegisterType((*ChaincodeLocator)(nil), "service.ChaincodeLocator")
+	proto.RegisterType((*ChaincodeExec)(nil), "service.ChaincodeExec")
+}
+
+func init() { proto.RegisterFile("chaincode.proto", fileDescriptor_97136ef4b384cc22) }
+
+var fileDescriptor_97136ef4b384cc22 = []byte{
+	// 424 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x52, 0x5d, 0x8b, 0x13, 0x31,
+	0x14, 0x75, 0xb6, 0x1f, 0xcb, 0xdc, 0x5d, 0x6b, 0xb9, 0xc8, 0x3a, 0x0e, 0x3e, 0x94, 0x3e, 0x68,
+	0x51, 0x9c, 0x4a, 0x7d, 0x59, 0x56, 0x45, 0x44, 0xe7, 0xa1, 0x2a, 0xea, 0x0e, 0xab, 0xe0, 0xd3,
+	0x92, 0x66, 0xaf, 0xed, 0xd0, 0x9a, 0x84, 0x24, 0x2d, 0xce, 0xbf, 0xf5, 0x8f, 0x08, 0x92, 0xcc,
+	0x74, 0xea, 0xb0, 0x28, 0xd2, 0xb7, 0x9b, 0xdc, 0x73, 0xce, 0x3d, 0x27, 0xb9, 0x70, 0x8b, 0x2f,
+	0x58, 0x2e, 0xb8, 0xbc, 0xa2, 0x44, 0x69, 0x69, 0x25, 0x1e, 0x1a, 0xd2, 0x9b, 0x9c, 0x53, 0xfc,
+	0x6a, 0x9e, 0xdb, 0xc5, 0x7a, 0x96, 0x70, 0xf9, 0x7d, 0xbc, 0x28, 0x14, 0xe9, 0x15, 0x5d, 0xcd,
+	0x49, 0x8f, 0xbf, 0xb1, 0x99, 0xce, 0xf9, 0xd8, 0xa3, 0xcd, 0x58, 0x11, 0x69, 0x57, 0x2b, 0x69,
+	0xd8, 0xea, 0x52, 0x93, 0x51, 0x52, 0x98, 0x4a, 0x2b, 0x7e, 0xf9, 0xff, 0x12, 0xb5, 0x8d, 0x4b,
+	0xda, 0x90, 0xb0, 0xa5, 0xc0, 0xf0, 0x67, 0x00, 0xbd, 0xd7, 0xdb, 0xce, 0x54, 0xa8, 0xb5, 0xc5,
+	0x7b, 0x10, 0xd6, 0xd8, 0x28, 0x18, 0x04, 0xa3, 0x30, 0xdb, 0x5d, 0x60, 0x04, 0x87, 0x7c, 0xc1,
+	0x84, 0xa0, 0x55, 0x74, 0xe0, 0x7b, 0xdb, 0x23, 0x22, 0xb4, 0x99, 0x9e, 0x9b, 0xa8, 0x35, 0x68,
+	0x8d, 0x8e, 0x33, 0x5f, 0xe3, 0x1b, 0x08, 0xad, 0x66, 0xc2, 0xe4, 0x24, 0x6c, 0xd4, 0x1e, 0xb4,
+	0x46, 0x47, 0x93, 0xfb, 0x49, 0x95, 0x3f, 0x69, 0xce, 0x4d, 0x2e, 0xb6, 0xc0, 0x54, 0x58, 0x5d,
+	0x64, 0x3b, 0x62, 0xfc, 0x1c, 0x7a, 0xcd, 0x26, 0xf6, 0xa1, 0xb5, 0xa4, 0xa2, 0x72, 0xe7, 0x4a,
+	0xbc, 0x0d, 0x9d, 0x0d, 0x5b, 0xad, 0xc9, 0xbb, 0x3a, 0xce, 0xca, 0xc3, 0xd9, 0xc1, 0x69, 0x30,
+	0x7c, 0x0b, 0xfd, 0x7a, 0xd2, 0x7b, 0xc9, 0x99, 0x95, 0x7a, 0xdf, 0x8c, 0xc3, 0x25, 0xdc, 0xac,
+	0xb5, 0xd2, 0x1f, 0xc4, 0xf1, 0x11, 0xb4, 0x6d, 0xa1, 0x4a, 0x8d, 0xde, 0xe4, 0x4e, 0x9d, 0x6d,
+	0x2a, 0x36, 0x6e, 0x54, 0x2e, 0xc5, 0x45, 0xa1, 0x28, 0xf3, 0x20, 0x7c, 0x0c, 0x9d, 0xdc, 0x45,
+	0xf5, 0xaa, 0x47, 0x7f, 0xa0, 0x9b, 0x2f, 0x91, 0x95, 0xa8, 0x87, 0x0f, 0xa0, 0xd7, 0x94, 0xc1,
+	0x10, 0x3a, 0xe7, 0x9f, 0xd3, 0xec, 0x6b, 0xff, 0x06, 0x02, 0x74, 0xa7, 0x1f, 0xbe, 0x7c, 0x7c,
+	0x97, 0xf6, 0x83, 0xc9, 0xaf, 0x00, 0xc2, 0x5a, 0x02, 0x4f, 0xa1, 0xed, 0xad, 0x9d, 0x5c, 0x97,
+	0x77, 0xf7, 0x71, 0x54, 0x7e, 0xbd, 0x49, 0x3e, 0x55, 0x4b, 0x95, 0x55, 0x3b, 0x85, 0x67, 0xd0,
+	0x39, 0x5f, 0x93, 0x2e, 0xf0, 0x6f, 0xce, 0xfe, 0xc1, 0x7d, 0x06, 0x5d, 0x67, 0x76, 0x49, 0xfb,
+	0x90, 0x5f, 0x40, 0x37, 0x75, 0x4b, 0x69, 0xf0, 0xee, 0x75, 0x72, 0xf5, 0x67, 0xf1, 0xc9, 0x96,
+	0xbe, 0x8b, 0xe3, 0x38, 0x4f, 0x82, 0x59, 0xd7, 0x37, 0x9e, 0xfe, 0x0e, 0x00, 0x00, 0xff, 0xff,
+	0x27, 0xc3, 0x64, 0xce, 0x6b, 0x03, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -121,12 +263,17 @@ var _ grpc.ClientConn
 // is compatible with the grpc package it is being compiled against.
 const _ = grpc.SupportPackageIsVersion4
 
-// Client API for Chaincode service
-
+// ChaincodeClient is the client API for Chaincode service.
+//
+// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
 type ChaincodeClient interface {
-	// Query chaincode on home peer
-	Query(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*protos.ProposalResponse, error)
-	Invoke(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*protos.ProposalResponse, error)
+	// Exec: Query or Invoke
+	Exec(ctx context.Context, in *ChaincodeExec, opts ...grpc.CallOption) (*peer.ProposalResponse, error)
+	// Query chaincode on home peer. Do NOT send to orderer.
+	Query(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*peer.ProposalResponse, error)
+	// Invoke chaincode on peers, according to endorsement policy and the SEND to orderer
+	Invoke(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*peer.ProposalResponse, error)
+	// Chaincode events stream
 	Events(ctx context.Context, in *ChaincodeLocator, opts ...grpc.CallOption) (Chaincode_EventsClient, error)
 }
 
@@ -138,18 +285,27 @@ func NewChaincodeClient(cc *grpc.ClientConn) ChaincodeClient {
 	return &chaincodeClient{cc}
 }
 
-func (c *chaincodeClient) Query(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*protos.ProposalResponse, error) {
-	out := new(protos.ProposalResponse)
-	err := grpc.Invoke(ctx, "/service.Chaincode/Query", in, out, c.cc, opts...)
+func (c *chaincodeClient) Exec(ctx context.Context, in *ChaincodeExec, opts ...grpc.CallOption) (*peer.ProposalResponse, error) {
+	out := new(peer.ProposalResponse)
+	err := c.cc.Invoke(ctx, "/service.Chaincode/Exec", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *chaincodeClient) Invoke(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*protos.ProposalResponse, error) {
-	out := new(protos.ProposalResponse)
-	err := grpc.Invoke(ctx, "/service.Chaincode/Invoke", in, out, c.cc, opts...)
+func (c *chaincodeClient) Query(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*peer.ProposalResponse, error) {
+	out := new(peer.ProposalResponse)
+	err := c.cc.Invoke(ctx, "/service.Chaincode/Query", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *chaincodeClient) Invoke(ctx context.Context, in *ChaincodeInput, opts ...grpc.CallOption) (*peer.ProposalResponse, error) {
+	out := new(peer.ProposalResponse)
+	err := c.cc.Invoke(ctx, "/service.Chaincode/Invoke", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -157,7 +313,7 @@ func (c *chaincodeClient) Invoke(ctx context.Context, in *ChaincodeInput, opts .
 }
 
 func (c *chaincodeClient) Events(ctx context.Context, in *ChaincodeLocator, opts ...grpc.CallOption) (Chaincode_EventsClient, error) {
-	stream, err := grpc.NewClientStream(ctx, &_Chaincode_serviceDesc.Streams[0], c.cc, "/service.Chaincode/Events", opts...)
+	stream, err := c.cc.NewStream(ctx, &_Chaincode_serviceDesc.Streams[0], "/service.Chaincode/Events", opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -172,7 +328,7 @@ func (c *chaincodeClient) Events(ctx context.Context, in *ChaincodeLocator, opts
 }
 
 type Chaincode_EventsClient interface {
-	Recv() (*protos1.ChaincodeEvent, error)
+	Recv() (*peer.ChaincodeEvent, error)
 	grpc.ClientStream
 }
 
@@ -180,25 +336,46 @@ type chaincodeEventsClient struct {
 	grpc.ClientStream
 }
 
-func (x *chaincodeEventsClient) Recv() (*protos1.ChaincodeEvent, error) {
-	m := new(protos1.ChaincodeEvent)
+func (x *chaincodeEventsClient) Recv() (*peer.ChaincodeEvent, error) {
+	m := new(peer.ChaincodeEvent)
 	if err := x.ClientStream.RecvMsg(m); err != nil {
 		return nil, err
 	}
 	return m, nil
 }
 
-// Server API for Chaincode service
-
+// ChaincodeServer is the server API for Chaincode service.
 type ChaincodeServer interface {
-	// Query chaincode on home peer
-	Query(context.Context, *ChaincodeInput) (*protos.ProposalResponse, error)
-	Invoke(context.Context, *ChaincodeInput) (*protos.ProposalResponse, error)
+	// Exec: Query or Invoke
+	Exec(context.Context, *ChaincodeExec) (*peer.ProposalResponse, error)
+	// Query chaincode on home peer. Do NOT send to orderer.
+	Query(context.Context, *ChaincodeInput) (*peer.ProposalResponse, error)
+	// Invoke chaincode on peers, according to endorsement policy and the SEND to orderer
+	Invoke(context.Context, *ChaincodeInput) (*peer.ProposalResponse, error)
+	// Chaincode events stream
 	Events(*ChaincodeLocator, Chaincode_EventsServer) error
 }
 
 func RegisterChaincodeServer(s *grpc.Server, srv ChaincodeServer) {
 	s.RegisterService(&_Chaincode_serviceDesc, srv)
+}
+
+func _Chaincode_Exec_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ChaincodeExec)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ChaincodeServer).Exec(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/service.Chaincode/Exec",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ChaincodeServer).Exec(ctx, req.(*ChaincodeExec))
+	}
+	return interceptor(ctx, in, info, handler)
 }
 
 func _Chaincode_Query_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
@@ -246,7 +423,7 @@ func _Chaincode_Events_Handler(srv interface{}, stream grpc.ServerStream) error 
 }
 
 type Chaincode_EventsServer interface {
-	Send(*protos1.ChaincodeEvent) error
+	Send(*peer.ChaincodeEvent) error
 	grpc.ServerStream
 }
 
@@ -254,7 +431,7 @@ type chaincodeEventsServer struct {
 	grpc.ServerStream
 }
 
-func (x *chaincodeEventsServer) Send(m *protos1.ChaincodeEvent) error {
+func (x *chaincodeEventsServer) Send(m *peer.ChaincodeEvent) error {
 	return x.ServerStream.SendMsg(m)
 }
 
@@ -262,6 +439,10 @@ var _Chaincode_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "service.Chaincode",
 	HandlerType: (*ChaincodeServer)(nil),
 	Methods: []grpc.MethodDesc{
+		{
+			MethodName: "Exec",
+			Handler:    _Chaincode_Exec_Handler,
+		},
 		{
 			MethodName: "Query",
 			Handler:    _Chaincode_Query_Handler,
@@ -279,31 +460,4 @@ var _Chaincode_serviceDesc = grpc.ServiceDesc{
 		},
 	},
 	Metadata: "chaincode.proto",
-}
-
-func init() { proto.RegisterFile("chaincode.proto", fileDescriptor0) }
-
-var fileDescriptor0 = []byte{
-	// 335 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x51, 0x4f, 0x4b, 0xfb, 0x40,
-	0x10, 0x25, 0x4d, 0xff, 0x90, 0xf9, 0x95, 0xfe, 0xca, 0x22, 0x1a, 0x83, 0x87, 0xd2, 0x83, 0xf4,
-	0x94, 0x48, 0xbd, 0x48, 0x55, 0x44, 0xb4, 0x87, 0x8a, 0x07, 0x0d, 0xde, 0xcb, 0x76, 0x3b, 0x36,
-	0xa1, 0x71, 0x37, 0xec, 0x6e, 0x02, 0xf9, 0x7c, 0x7e, 0x11, 0x3f, 0x8a, 0x34, 0xdd, 0xa4, 0x14,
-	0x11, 0xa4, 0xb7, 0x99, 0x7d, 0xf3, 0xde, 0xcc, 0xdb, 0x07, 0xff, 0x59, 0x44, 0x63, 0xce, 0xc4,
-	0x12, 0xfd, 0x54, 0x0a, 0x2d, 0x48, 0x47, 0xa1, 0xcc, 0x63, 0x86, 0xde, 0xfd, 0x2a, 0xd6, 0x51,
-	0xb6, 0xf0, 0x99, 0xf8, 0x08, 0xa2, 0x22, 0x45, 0x99, 0xe0, 0x72, 0x85, 0x32, 0x78, 0xa7, 0x0b,
-	0x19, 0xb3, 0xa0, 0x9c, 0x56, 0x41, 0x8a, 0x28, 0x37, 0x75, 0x2a, 0x14, 0x4d, 0xe6, 0x12, 0x55,
-	0x2a, 0xb8, 0x32, 0x5a, 0xde, 0xdd, 0xdf, 0x25, 0xea, 0x33, 0xe6, 0x98, 0x23, 0xd7, 0x5b, 0x81,
-	0xe1, 0x97, 0x05, 0xbd, 0x87, 0x0a, 0x99, 0xf1, 0x34, 0xd3, 0xe4, 0x0c, 0x9c, 0x7a, 0xd6, 0xb5,
-	0x06, 0xd6, 0xc8, 0x09, 0x77, 0x0f, 0xc4, 0x85, 0x0e, 0x8b, 0x28, 0xe7, 0x98, 0xb8, 0x8d, 0x12,
-	0xab, 0x5a, 0x42, 0xa0, 0x49, 0xe5, 0x4a, 0xb9, 0xf6, 0xc0, 0x1e, 0x75, 0xc3, 0xb2, 0x26, 0x8f,
-	0xe0, 0x68, 0x49, 0xb9, 0x8a, 0x91, 0x6b, 0xb7, 0x39, 0xb0, 0x47, 0xff, 0xc6, 0xe7, 0xbe, 0xf1,
-	0xef, 0xef, 0xef, 0xf5, 0xdf, 0xaa, 0xc1, 0x29, 0xd7, 0xb2, 0x08, 0x77, 0x44, 0xef, 0x06, 0x7a,
-	0xfb, 0x20, 0xe9, 0x83, 0xbd, 0xc6, 0xc2, 0x5c, 0xb7, 0x29, 0xc9, 0x11, 0xb4, 0x72, 0x9a, 0x64,
-	0x58, 0x5e, 0xd5, 0x0d, 0xb7, 0xcd, 0xa4, 0x71, 0x65, 0x0d, 0x9f, 0xa0, 0x5f, 0x6f, 0x7a, 0x16,
-	0x8c, 0x6a, 0x21, 0x0f, 0xf5, 0x38, 0xfe, 0xb4, 0xc0, 0xa9, 0xc5, 0xc8, 0x04, 0x5a, 0xaf, 0x19,
-	0xca, 0x82, 0x9c, 0xfc, 0xe2, 0xc9, 0x73, 0xb7, 0xdf, 0xac, 0xfc, 0x17, 0x13, 0x60, 0x68, 0xf2,
-	0x23, 0xd7, 0xd0, 0x9e, 0xf1, 0x5c, 0xac, 0xf1, 0x10, 0xf2, 0x2d, 0xb4, 0xa7, 0x9b, 0x10, 0x15,
-	0x39, 0xfd, 0x49, 0x36, 0x1e, 0xbd, 0xe3, 0x8a, 0x5e, 0x23, 0x25, 0xe7, 0xc2, 0x5a, 0xb4, 0x4b,
-	0xe0, 0xf2, 0x3b, 0x00, 0x00, 0xff, 0xff, 0x63, 0x4d, 0xd9, 0x47, 0x9b, 0x02, 0x00, 0x00,
 }
