@@ -2,18 +2,20 @@ package cpaper_asservice_test
 
 import (
 	"crypto/rand"
-	"path"
+	"io/ioutil"
 	"testing"
 	"time"
 
-	"github.com/golang/protobuf/ptypes"
-	"github.com/hyperledger/fabric/protos/peer"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
+	"github.com/golang/protobuf/ptypes"
+	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/hyperledger/fabric/msp"
+	"github.com/hyperledger/fabric/protos/peer"
 	"github.com/s7techlab/cckit/examples/cpaper_asservice"
 	"github.com/s7techlab/cckit/examples/cpaper_asservice/schema"
 	s "github.com/s7techlab/cckit/examples/cpaper_asservice/service"
-	"github.com/s7techlab/cckit/examples/cpaper_asservice/testdata"
 	"github.com/s7techlab/cckit/extensions/encryption"
 	"github.com/s7techlab/cckit/router"
 	testcc "github.com/s7techlab/cckit/testing"
@@ -48,6 +50,8 @@ var (
 		FaceValue:    100000,
 		ExternalId:   "EXT0001",
 	}
+
+	identity msp.SigningIdentity
 )
 
 var _ = Describe(`CommercialPaper`, func() {
@@ -69,9 +73,14 @@ var _ = Describe(`CommercialPaper`, func() {
 		// all queries/invokes arguments to cc will be encrypted
 		ccEncWrapped = encryption.NewMockStub(ccEnc, encKey)
 
+		identity, err = testcc.IdentityFromFile(MspName, `./testdata/admin.pem`, ioutil.ReadFile)
+		Expect(err).NotTo(HaveOccurred())
 		// Init chaincode with admin identity
 		expectcc.ResponseOk(
-			cc.From(testdata.GetTestIdentity(MspName, path.Join("testdata", "admin", "admin.pem"))).Init())
+			cc.From(identity).Init())
+
+		//js, _ := json.Marshal(issuePayload)
+		//fmt.Println(string(js))
 	})
 
 	Describe("Commercial Paper lifecycle", func() {
@@ -118,7 +127,7 @@ var _ = Describe(`CommercialPaper`, func() {
 		})
 
 		It("Allow issuer to get a list of commercial papers", func() {
-			queryResponse := cc.Query(s.CPaperChaincode_List)
+			queryResponse := cc.Query(s.CPaperChaincode_List, &empty.Empty{})
 
 			papers := expectcc.PayloadIs(queryResponse, &schema.CommercialPaperList{}).(*schema.CommercialPaperList)
 
@@ -194,7 +203,7 @@ var _ = Describe(`CommercialPaper`, func() {
 			}))
 
 			// Validate there are 0 Commercial Papers in the world state
-			queryResponse := cc.Query(s.CPaperChaincode_List)
+			queryResponse := cc.Query(s.CPaperChaincode_List, &empty.Empty{})
 			papers := expectcc.PayloadIs(queryResponse, &schema.CommercialPaperList{}).(*schema.CommercialPaperList)
 
 			Expect(len(papers.Items)).To(BeNumerically("==", 0))
