@@ -1,28 +1,15 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // source: schema.proto
 
-/*
-Package schema is a generated protocol buffer package.
-
-It is generated from these files:
-	schema.proto
-
-It has these top-level messages:
-	CommercialPaper
-	CommercialPaperId
-	ExternalId
-	CommercialPaperList
-	IssueCommercialPaper
-	BuyCommercialPaper
-	RedeemCommercialPaper
-*/
 package schema
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
-import google_protobuf "github.com/golang/protobuf/ptypes/timestamp"
-import _ "github.com/mwitkow/go-proto-validators"
+import (
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	timestamp "github.com/golang/protobuf/ptypes/timestamp"
+	_ "github.com/mwitkow/go-proto-validators"
+	math "math"
+)
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
@@ -33,7 +20,7 @@ var _ = math.Inf
 // is compatible with the proto package it is being compiled against.
 // A compilation error at this line likely means your copy of the
 // proto package needs to be updated.
-const _ = proto.ProtoPackageIsVersion2 // please upgrade the proto package
+const _ = proto.ProtoPackageIsVersion3 // please upgrade the proto package
 
 type CommercialPaper_State int32
 
@@ -48,6 +35,7 @@ var CommercialPaper_State_name = map[int32]string{
 	1: "TRADING",
 	2: "REDEEMED",
 }
+
 var CommercialPaper_State_value = map[string]int32{
 	"ISSUED":   0,
 	"TRADING":  1,
@@ -57,26 +45,52 @@ var CommercialPaper_State_value = map[string]int32{
 func (x CommercialPaper_State) String() string {
 	return proto.EnumName(CommercialPaper_State_name, int32(x))
 }
-func (CommercialPaper_State) EnumDescriptor() ([]byte, []int) { return fileDescriptor0, []int{0, 0} }
+
+func (CommercialPaper_State) EnumDescriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{0, 0}
+}
 
 // Commercial Paper state entry
 type CommercialPaper struct {
 	// Issuer and Paper number comprises composite primary key of Commercial paper entry
-	Issuer       string                     `protobuf:"bytes,1,opt,name=issuer" json:"issuer,omitempty"`
-	PaperNumber  string                     `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber" json:"paper_number,omitempty"`
-	Owner        string                     `protobuf:"bytes,3,opt,name=owner" json:"owner,omitempty"`
-	IssueDate    *google_protobuf.Timestamp `protobuf:"bytes,4,opt,name=issue_date,json=issueDate" json:"issue_date,omitempty"`
-	MaturityDate *google_protobuf.Timestamp `protobuf:"bytes,5,opt,name=maturity_date,json=maturityDate" json:"maturity_date,omitempty"`
-	FaceValue    int32                      `protobuf:"varint,6,opt,name=face_value,json=faceValue" json:"face_value,omitempty"`
-	State        CommercialPaper_State      `protobuf:"varint,7,opt,name=state,enum=schema.CommercialPaper_State" json:"state,omitempty"`
+	Issuer       string                `protobuf:"bytes,1,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	PaperNumber  string                `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber,proto3" json:"paper_number,omitempty"`
+	Owner        string                `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`
+	IssueDate    *timestamp.Timestamp  `protobuf:"bytes,4,opt,name=issue_date,json=issueDate,proto3" json:"issue_date,omitempty"`
+	MaturityDate *timestamp.Timestamp  `protobuf:"bytes,5,opt,name=maturity_date,json=maturityDate,proto3" json:"maturity_date,omitempty"`
+	FaceValue    int32                 `protobuf:"varint,6,opt,name=face_value,json=faceValue,proto3" json:"face_value,omitempty"`
+	State        CommercialPaper_State `protobuf:"varint,7,opt,name=state,proto3,enum=schema.CommercialPaper_State" json:"state,omitempty"`
 	// Additional unique field for entry
-	ExternalId string `protobuf:"bytes,8,opt,name=external_id,json=externalId" json:"external_id,omitempty"`
+	ExternalId           string   `protobuf:"bytes,8,opt,name=external_id,json=externalId,proto3" json:"external_id,omitempty"`
+	XXX_NoUnkeyedLiteral struct{} `json:"-"`
+	XXX_unrecognized     []byte   `json:"-"`
+	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *CommercialPaper) Reset()                    { *m = CommercialPaper{} }
-func (m *CommercialPaper) String() string            { return proto.CompactTextString(m) }
-func (*CommercialPaper) ProtoMessage()               {}
-func (*CommercialPaper) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{0} }
+func (m *CommercialPaper) Reset()         { *m = CommercialPaper{} }
+func (m *CommercialPaper) String() string { return proto.CompactTextString(m) }
+func (*CommercialPaper) ProtoMessage()    {}
+func (*CommercialPaper) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{0}
+}
+
+func (m *CommercialPaper) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_CommercialPaper.Unmarshal(m, b)
+}
+func (m *CommercialPaper) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_CommercialPaper.Marshal(b, m, deterministic)
+}
+func (m *CommercialPaper) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_CommercialPaper.Merge(m, src)
+}
+func (m *CommercialPaper) XXX_Size() int {
+	return xxx_messageInfo_CommercialPaper.Size(m)
+}
+func (m *CommercialPaper) XXX_DiscardUnknown() {
+	xxx_messageInfo_CommercialPaper.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_CommercialPaper proto.InternalMessageInfo
 
 func (m *CommercialPaper) GetIssuer() string {
 	if m != nil {
@@ -99,14 +113,14 @@ func (m *CommercialPaper) GetOwner() string {
 	return ""
 }
 
-func (m *CommercialPaper) GetIssueDate() *google_protobuf.Timestamp {
+func (m *CommercialPaper) GetIssueDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.IssueDate
 	}
 	return nil
 }
 
-func (m *CommercialPaper) GetMaturityDate() *google_protobuf.Timestamp {
+func (m *CommercialPaper) GetMaturityDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.MaturityDate
 	}
@@ -136,14 +150,37 @@ func (m *CommercialPaper) GetExternalId() string {
 
 // CommercialPaperId identifier part
 type CommercialPaperId struct {
-	Issuer      string `protobuf:"bytes,1,opt,name=issuer" json:"issuer,omitempty"`
-	PaperNumber string `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber" json:"paper_number,omitempty"`
+	Issuer               string   `protobuf:"bytes,1,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	PaperNumber          string   `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber,proto3" json:"paper_number,omitempty"`
+	XXX_NoUnkeyedLiteral struct{} `json:"-"`
+	XXX_unrecognized     []byte   `json:"-"`
+	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *CommercialPaperId) Reset()                    { *m = CommercialPaperId{} }
-func (m *CommercialPaperId) String() string            { return proto.CompactTextString(m) }
-func (*CommercialPaperId) ProtoMessage()               {}
-func (*CommercialPaperId) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{1} }
+func (m *CommercialPaperId) Reset()         { *m = CommercialPaperId{} }
+func (m *CommercialPaperId) String() string { return proto.CompactTextString(m) }
+func (*CommercialPaperId) ProtoMessage()    {}
+func (*CommercialPaperId) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{1}
+}
+
+func (m *CommercialPaperId) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_CommercialPaperId.Unmarshal(m, b)
+}
+func (m *CommercialPaperId) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_CommercialPaperId.Marshal(b, m, deterministic)
+}
+func (m *CommercialPaperId) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_CommercialPaperId.Merge(m, src)
+}
+func (m *CommercialPaperId) XXX_Size() int {
+	return xxx_messageInfo_CommercialPaperId.Size(m)
+}
+func (m *CommercialPaperId) XXX_DiscardUnknown() {
+	xxx_messageInfo_CommercialPaperId.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_CommercialPaperId proto.InternalMessageInfo
 
 func (m *CommercialPaperId) GetIssuer() string {
 	if m != nil {
@@ -161,13 +198,36 @@ func (m *CommercialPaperId) GetPaperNumber() string {
 
 // ExternalId
 type ExternalId struct {
-	Id string `protobuf:"bytes,1,opt,name=id" json:"id,omitempty"`
+	Id                   string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	XXX_NoUnkeyedLiteral struct{} `json:"-"`
+	XXX_unrecognized     []byte   `json:"-"`
+	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *ExternalId) Reset()                    { *m = ExternalId{} }
-func (m *ExternalId) String() string            { return proto.CompactTextString(m) }
-func (*ExternalId) ProtoMessage()               {}
-func (*ExternalId) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{2} }
+func (m *ExternalId) Reset()         { *m = ExternalId{} }
+func (m *ExternalId) String() string { return proto.CompactTextString(m) }
+func (*ExternalId) ProtoMessage()    {}
+func (*ExternalId) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{2}
+}
+
+func (m *ExternalId) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_ExternalId.Unmarshal(m, b)
+}
+func (m *ExternalId) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_ExternalId.Marshal(b, m, deterministic)
+}
+func (m *ExternalId) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_ExternalId.Merge(m, src)
+}
+func (m *ExternalId) XXX_Size() int {
+	return xxx_messageInfo_ExternalId.Size(m)
+}
+func (m *ExternalId) XXX_DiscardUnknown() {
+	xxx_messageInfo_ExternalId.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_ExternalId proto.InternalMessageInfo
 
 func (m *ExternalId) GetId() string {
 	if m != nil {
@@ -178,13 +238,36 @@ func (m *ExternalId) GetId() string {
 
 // Container for returning multiple entities
 type CommercialPaperList struct {
-	Items []*CommercialPaper `protobuf:"bytes,1,rep,name=items" json:"items,omitempty"`
+	Items                []*CommercialPaper `protobuf:"bytes,1,rep,name=items,proto3" json:"items,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}           `json:"-"`
+	XXX_unrecognized     []byte             `json:"-"`
+	XXX_sizecache        int32              `json:"-"`
 }
 
-func (m *CommercialPaperList) Reset()                    { *m = CommercialPaperList{} }
-func (m *CommercialPaperList) String() string            { return proto.CompactTextString(m) }
-func (*CommercialPaperList) ProtoMessage()               {}
-func (*CommercialPaperList) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{3} }
+func (m *CommercialPaperList) Reset()         { *m = CommercialPaperList{} }
+func (m *CommercialPaperList) String() string { return proto.CompactTextString(m) }
+func (*CommercialPaperList) ProtoMessage()    {}
+func (*CommercialPaperList) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{3}
+}
+
+func (m *CommercialPaperList) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_CommercialPaperList.Unmarshal(m, b)
+}
+func (m *CommercialPaperList) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_CommercialPaperList.Marshal(b, m, deterministic)
+}
+func (m *CommercialPaperList) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_CommercialPaperList.Merge(m, src)
+}
+func (m *CommercialPaperList) XXX_Size() int {
+	return xxx_messageInfo_CommercialPaperList.Size(m)
+}
+func (m *CommercialPaperList) XXX_DiscardUnknown() {
+	xxx_messageInfo_CommercialPaperList.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_CommercialPaperList proto.InternalMessageInfo
 
 func (m *CommercialPaperList) GetItems() []*CommercialPaper {
 	if m != nil {
@@ -195,19 +278,42 @@ func (m *CommercialPaperList) GetItems() []*CommercialPaper {
 
 // IssueCommercialPaper event
 type IssueCommercialPaper struct {
-	Issuer       string                     `protobuf:"bytes,1,opt,name=issuer" json:"issuer,omitempty"`
-	PaperNumber  string                     `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber" json:"paper_number,omitempty"`
-	IssueDate    *google_protobuf.Timestamp `protobuf:"bytes,3,opt,name=issue_date,json=issueDate" json:"issue_date,omitempty"`
-	MaturityDate *google_protobuf.Timestamp `protobuf:"bytes,4,opt,name=maturity_date,json=maturityDate" json:"maturity_date,omitempty"`
-	FaceValue    int32                      `protobuf:"varint,5,opt,name=face_value,json=faceValue" json:"face_value,omitempty"`
+	Issuer       string               `protobuf:"bytes,1,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	PaperNumber  string               `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber,proto3" json:"paper_number,omitempty"`
+	IssueDate    *timestamp.Timestamp `protobuf:"bytes,3,opt,name=issue_date,json=issueDate,proto3" json:"issue_date,omitempty"`
+	MaturityDate *timestamp.Timestamp `protobuf:"bytes,4,opt,name=maturity_date,json=maturityDate,proto3" json:"maturity_date,omitempty"`
+	FaceValue    int32                `protobuf:"varint,5,opt,name=face_value,json=faceValue,proto3" json:"face_value,omitempty"`
 	// external_id  - once more uniq id of state entry
-	ExternalId string `protobuf:"bytes,6,opt,name=external_id,json=externalId" json:"external_id,omitempty"`
+	ExternalId           string   `protobuf:"bytes,6,opt,name=external_id,json=externalId,proto3" json:"external_id,omitempty"`
+	XXX_NoUnkeyedLiteral struct{} `json:"-"`
+	XXX_unrecognized     []byte   `json:"-"`
+	XXX_sizecache        int32    `json:"-"`
 }
 
-func (m *IssueCommercialPaper) Reset()                    { *m = IssueCommercialPaper{} }
-func (m *IssueCommercialPaper) String() string            { return proto.CompactTextString(m) }
-func (*IssueCommercialPaper) ProtoMessage()               {}
-func (*IssueCommercialPaper) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{4} }
+func (m *IssueCommercialPaper) Reset()         { *m = IssueCommercialPaper{} }
+func (m *IssueCommercialPaper) String() string { return proto.CompactTextString(m) }
+func (*IssueCommercialPaper) ProtoMessage()    {}
+func (*IssueCommercialPaper) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{4}
+}
+
+func (m *IssueCommercialPaper) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_IssueCommercialPaper.Unmarshal(m, b)
+}
+func (m *IssueCommercialPaper) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_IssueCommercialPaper.Marshal(b, m, deterministic)
+}
+func (m *IssueCommercialPaper) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_IssueCommercialPaper.Merge(m, src)
+}
+func (m *IssueCommercialPaper) XXX_Size() int {
+	return xxx_messageInfo_IssueCommercialPaper.Size(m)
+}
+func (m *IssueCommercialPaper) XXX_DiscardUnknown() {
+	xxx_messageInfo_IssueCommercialPaper.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_IssueCommercialPaper proto.InternalMessageInfo
 
 func (m *IssueCommercialPaper) GetIssuer() string {
 	if m != nil {
@@ -223,14 +329,14 @@ func (m *IssueCommercialPaper) GetPaperNumber() string {
 	return ""
 }
 
-func (m *IssueCommercialPaper) GetIssueDate() *google_protobuf.Timestamp {
+func (m *IssueCommercialPaper) GetIssueDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.IssueDate
 	}
 	return nil
 }
 
-func (m *IssueCommercialPaper) GetMaturityDate() *google_protobuf.Timestamp {
+func (m *IssueCommercialPaper) GetMaturityDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.MaturityDate
 	}
@@ -253,18 +359,41 @@ func (m *IssueCommercialPaper) GetExternalId() string {
 
 // BuyCommercialPaper event
 type BuyCommercialPaper struct {
-	Issuer       string                     `protobuf:"bytes,1,opt,name=issuer" json:"issuer,omitempty"`
-	PaperNumber  string                     `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber" json:"paper_number,omitempty"`
-	CurrentOwner string                     `protobuf:"bytes,3,opt,name=current_owner,json=currentOwner" json:"current_owner,omitempty"`
-	NewOwner     string                     `protobuf:"bytes,4,opt,name=new_owner,json=newOwner" json:"new_owner,omitempty"`
-	Price        int32                      `protobuf:"varint,5,opt,name=price" json:"price,omitempty"`
-	PurchaseDate *google_protobuf.Timestamp `protobuf:"bytes,6,opt,name=purchase_date,json=purchaseDate" json:"purchase_date,omitempty"`
+	Issuer               string               `protobuf:"bytes,1,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	PaperNumber          string               `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber,proto3" json:"paper_number,omitempty"`
+	CurrentOwner         string               `protobuf:"bytes,3,opt,name=current_owner,json=currentOwner,proto3" json:"current_owner,omitempty"`
+	NewOwner             string               `protobuf:"bytes,4,opt,name=new_owner,json=newOwner,proto3" json:"new_owner,omitempty"`
+	Price                int32                `protobuf:"varint,5,opt,name=price,proto3" json:"price,omitempty"`
+	PurchaseDate         *timestamp.Timestamp `protobuf:"bytes,6,opt,name=purchase_date,json=purchaseDate,proto3" json:"purchase_date,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
+	XXX_unrecognized     []byte               `json:"-"`
+	XXX_sizecache        int32                `json:"-"`
 }
 
-func (m *BuyCommercialPaper) Reset()                    { *m = BuyCommercialPaper{} }
-func (m *BuyCommercialPaper) String() string            { return proto.CompactTextString(m) }
-func (*BuyCommercialPaper) ProtoMessage()               {}
-func (*BuyCommercialPaper) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{5} }
+func (m *BuyCommercialPaper) Reset()         { *m = BuyCommercialPaper{} }
+func (m *BuyCommercialPaper) String() string { return proto.CompactTextString(m) }
+func (*BuyCommercialPaper) ProtoMessage()    {}
+func (*BuyCommercialPaper) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{5}
+}
+
+func (m *BuyCommercialPaper) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_BuyCommercialPaper.Unmarshal(m, b)
+}
+func (m *BuyCommercialPaper) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_BuyCommercialPaper.Marshal(b, m, deterministic)
+}
+func (m *BuyCommercialPaper) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_BuyCommercialPaper.Merge(m, src)
+}
+func (m *BuyCommercialPaper) XXX_Size() int {
+	return xxx_messageInfo_BuyCommercialPaper.Size(m)
+}
+func (m *BuyCommercialPaper) XXX_DiscardUnknown() {
+	xxx_messageInfo_BuyCommercialPaper.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_BuyCommercialPaper proto.InternalMessageInfo
 
 func (m *BuyCommercialPaper) GetIssuer() string {
 	if m != nil {
@@ -301,7 +430,7 @@ func (m *BuyCommercialPaper) GetPrice() int32 {
 	return 0
 }
 
-func (m *BuyCommercialPaper) GetPurchaseDate() *google_protobuf.Timestamp {
+func (m *BuyCommercialPaper) GetPurchaseDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.PurchaseDate
 	}
@@ -310,16 +439,39 @@ func (m *BuyCommercialPaper) GetPurchaseDate() *google_protobuf.Timestamp {
 
 // RedeemCommercialPaper event
 type RedeemCommercialPaper struct {
-	Issuer         string                     `protobuf:"bytes,1,opt,name=issuer" json:"issuer,omitempty"`
-	PaperNumber    string                     `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber" json:"paper_number,omitempty"`
-	RedeemingOwner string                     `protobuf:"bytes,3,opt,name=redeeming_owner,json=redeemingOwner" json:"redeeming_owner,omitempty"`
-	RedeemDate     *google_protobuf.Timestamp `protobuf:"bytes,4,opt,name=redeem_date,json=redeemDate" json:"redeem_date,omitempty"`
+	Issuer               string               `protobuf:"bytes,1,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	PaperNumber          string               `protobuf:"bytes,2,opt,name=paper_number,json=paperNumber,proto3" json:"paper_number,omitempty"`
+	RedeemingOwner       string               `protobuf:"bytes,3,opt,name=redeeming_owner,json=redeemingOwner,proto3" json:"redeeming_owner,omitempty"`
+	RedeemDate           *timestamp.Timestamp `protobuf:"bytes,4,opt,name=redeem_date,json=redeemDate,proto3" json:"redeem_date,omitempty"`
+	XXX_NoUnkeyedLiteral struct{}             `json:"-"`
+	XXX_unrecognized     []byte               `json:"-"`
+	XXX_sizecache        int32                `json:"-"`
 }
 
-func (m *RedeemCommercialPaper) Reset()                    { *m = RedeemCommercialPaper{} }
-func (m *RedeemCommercialPaper) String() string            { return proto.CompactTextString(m) }
-func (*RedeemCommercialPaper) ProtoMessage()               {}
-func (*RedeemCommercialPaper) Descriptor() ([]byte, []int) { return fileDescriptor0, []int{6} }
+func (m *RedeemCommercialPaper) Reset()         { *m = RedeemCommercialPaper{} }
+func (m *RedeemCommercialPaper) String() string { return proto.CompactTextString(m) }
+func (*RedeemCommercialPaper) ProtoMessage()    {}
+func (*RedeemCommercialPaper) Descriptor() ([]byte, []int) {
+	return fileDescriptor_1c5fb4d8cc22d66a, []int{6}
+}
+
+func (m *RedeemCommercialPaper) XXX_Unmarshal(b []byte) error {
+	return xxx_messageInfo_RedeemCommercialPaper.Unmarshal(m, b)
+}
+func (m *RedeemCommercialPaper) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	return xxx_messageInfo_RedeemCommercialPaper.Marshal(b, m, deterministic)
+}
+func (m *RedeemCommercialPaper) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_RedeemCommercialPaper.Merge(m, src)
+}
+func (m *RedeemCommercialPaper) XXX_Size() int {
+	return xxx_messageInfo_RedeemCommercialPaper.Size(m)
+}
+func (m *RedeemCommercialPaper) XXX_DiscardUnknown() {
+	xxx_messageInfo_RedeemCommercialPaper.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_RedeemCommercialPaper proto.InternalMessageInfo
 
 func (m *RedeemCommercialPaper) GetIssuer() string {
 	if m != nil {
@@ -342,7 +494,7 @@ func (m *RedeemCommercialPaper) GetRedeemingOwner() string {
 	return ""
 }
 
-func (m *RedeemCommercialPaper) GetRedeemDate() *google_protobuf.Timestamp {
+func (m *RedeemCommercialPaper) GetRedeemDate() *timestamp.Timestamp {
 	if m != nil {
 		return m.RedeemDate
 	}
@@ -350,6 +502,7 @@ func (m *RedeemCommercialPaper) GetRedeemDate() *google_protobuf.Timestamp {
 }
 
 func init() {
+	proto.RegisterEnum("schema.CommercialPaper_State", CommercialPaper_State_name, CommercialPaper_State_value)
 	proto.RegisterType((*CommercialPaper)(nil), "schema.CommercialPaper")
 	proto.RegisterType((*CommercialPaperId)(nil), "schema.CommercialPaperId")
 	proto.RegisterType((*ExternalId)(nil), "schema.ExternalId")
@@ -357,12 +510,11 @@ func init() {
 	proto.RegisterType((*IssueCommercialPaper)(nil), "schema.IssueCommercialPaper")
 	proto.RegisterType((*BuyCommercialPaper)(nil), "schema.BuyCommercialPaper")
 	proto.RegisterType((*RedeemCommercialPaper)(nil), "schema.RedeemCommercialPaper")
-	proto.RegisterEnum("schema.CommercialPaper_State", CommercialPaper_State_name, CommercialPaper_State_value)
 }
 
-func init() { proto.RegisterFile("schema.proto", fileDescriptor0) }
+func init() { proto.RegisterFile("schema.proto", fileDescriptor_1c5fb4d8cc22d66a) }
 
-var fileDescriptor0 = []byte{
+var fileDescriptor_1c5fb4d8cc22d66a = []byte{
 	// 597 bytes of a gzipped FileDescriptorProto
 	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xb4, 0x53, 0x5d, 0x6b, 0xd4, 0x40,
 	0x14, 0x6d, 0xb2, 0x4d, 0xda, 0xbd, 0xbb, 0x6d, 0xd7, 0xb1, 0x6a, 0x28, 0xad, 0x8d, 0x11, 0x71,
