@@ -1,29 +1,16 @@
 // Code generated by protoc-gen-gogo. DO NOT EDIT.
 // source: schema.proto
 
-/*
-Package schema is a generated protocol buffer package.
-
-It is generated from these files:
-	schema.proto
-
-It has these top-level messages:
-	CommercialPaper
-	CommercialPaperId
-	ExternalId
-	CommercialPaperList
-	IssueCommercialPaper
-	BuyCommercialPaper
-	RedeemCommercialPaper
-*/
 package schema
 
-import fmt "fmt"
-import go_proto_validators "github.com/mwitkow/go-proto-validators"
-import proto "github.com/golang/protobuf/proto"
-import math "math"
-import _ "github.com/golang/protobuf/ptypes/timestamp"
-import _ "github.com/mwitkow/go-proto-validators"
+import (
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	_ "github.com/golang/protobuf/ptypes/timestamp"
+	_ "github.com/mwitkow/go-proto-validators"
+	github_com_mwitkow_go_proto_validators "github.com/mwitkow/go-proto-validators"
+	math "math"
+)
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
@@ -32,13 +19,13 @@ var _ = math.Inf
 
 func (this *CommercialPaper) Validate() error {
 	if this.IssueDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.IssueDate); err != nil {
-			return go_proto_validators.FieldError("IssueDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.IssueDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("IssueDate", err)
 		}
 	}
 	if this.MaturityDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.MaturityDate); err != nil {
-			return go_proto_validators.FieldError("MaturityDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.MaturityDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("MaturityDate", err)
 		}
 	}
 	return nil
@@ -52,8 +39,8 @@ func (this *ExternalId) Validate() error {
 func (this *CommercialPaperList) Validate() error {
 	for _, item := range this.Items {
 		if item != nil {
-			if err := go_proto_validators.CallValidatorIfExists(item); err != nil {
-				return go_proto_validators.FieldError("Items", err)
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(item); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("Items", err)
 			}
 		}
 	}
@@ -61,77 +48,77 @@ func (this *CommercialPaperList) Validate() error {
 }
 func (this *IssueCommercialPaper) Validate() error {
 	if this.Issuer == "" {
-		return go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
+		return github_com_mwitkow_go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
 	}
 	if this.PaperNumber == "" {
-		return go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
+		return github_com_mwitkow_go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
 	}
 	if nil == this.IssueDate {
-		return go_proto_validators.FieldError("IssueDate", fmt.Errorf("message must exist"))
+		return github_com_mwitkow_go_proto_validators.FieldError("IssueDate", fmt.Errorf("message must exist"))
 	}
 	if this.IssueDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.IssueDate); err != nil {
-			return go_proto_validators.FieldError("IssueDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.IssueDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("IssueDate", err)
 		}
 	}
 	if nil == this.MaturityDate {
-		return go_proto_validators.FieldError("MaturityDate", fmt.Errorf("message must exist"))
+		return github_com_mwitkow_go_proto_validators.FieldError("MaturityDate", fmt.Errorf("message must exist"))
 	}
 	if this.MaturityDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.MaturityDate); err != nil {
-			return go_proto_validators.FieldError("MaturityDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.MaturityDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("MaturityDate", err)
 		}
 	}
 	if !(this.FaceValue > 0) {
-		return go_proto_validators.FieldError("FaceValue", fmt.Errorf(`value '%v' must be greater than '0'`, this.FaceValue))
+		return github_com_mwitkow_go_proto_validators.FieldError("FaceValue", fmt.Errorf(`value '%v' must be greater than '0'`, this.FaceValue))
 	}
 	if this.ExternalId == "" {
-		return go_proto_validators.FieldError("ExternalId", fmt.Errorf(`value '%v' must not be an empty string`, this.ExternalId))
+		return github_com_mwitkow_go_proto_validators.FieldError("ExternalId", fmt.Errorf(`value '%v' must not be an empty string`, this.ExternalId))
 	}
 	return nil
 }
 func (this *BuyCommercialPaper) Validate() error {
 	if this.Issuer == "" {
-		return go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
+		return github_com_mwitkow_go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
 	}
 	if this.PaperNumber == "" {
-		return go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
+		return github_com_mwitkow_go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
 	}
 	if this.CurrentOwner == "" {
-		return go_proto_validators.FieldError("CurrentOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.CurrentOwner))
+		return github_com_mwitkow_go_proto_validators.FieldError("CurrentOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.CurrentOwner))
 	}
 	if this.NewOwner == "" {
-		return go_proto_validators.FieldError("NewOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.NewOwner))
+		return github_com_mwitkow_go_proto_validators.FieldError("NewOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.NewOwner))
 	}
 	if !(this.Price > 0) {
-		return go_proto_validators.FieldError("Price", fmt.Errorf(`value '%v' must be greater than '0'`, this.Price))
+		return github_com_mwitkow_go_proto_validators.FieldError("Price", fmt.Errorf(`value '%v' must be greater than '0'`, this.Price))
 	}
 	if nil == this.PurchaseDate {
-		return go_proto_validators.FieldError("PurchaseDate", fmt.Errorf("message must exist"))
+		return github_com_mwitkow_go_proto_validators.FieldError("PurchaseDate", fmt.Errorf("message must exist"))
 	}
 	if this.PurchaseDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.PurchaseDate); err != nil {
-			return go_proto_validators.FieldError("PurchaseDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.PurchaseDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("PurchaseDate", err)
 		}
 	}
 	return nil
 }
 func (this *RedeemCommercialPaper) Validate() error {
 	if this.Issuer == "" {
-		return go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
+		return github_com_mwitkow_go_proto_validators.FieldError("Issuer", fmt.Errorf(`value '%v' must not be an empty string`, this.Issuer))
 	}
 	if this.PaperNumber == "" {
-		return go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
+		return github_com_mwitkow_go_proto_validators.FieldError("PaperNumber", fmt.Errorf(`value '%v' must not be an empty string`, this.PaperNumber))
 	}
 	if this.RedeemingOwner == "" {
-		return go_proto_validators.FieldError("RedeemingOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.RedeemingOwner))
+		return github_com_mwitkow_go_proto_validators.FieldError("RedeemingOwner", fmt.Errorf(`value '%v' must not be an empty string`, this.RedeemingOwner))
 	}
 	if nil == this.RedeemDate {
-		return go_proto_validators.FieldError("RedeemDate", fmt.Errorf("message must exist"))
+		return github_com_mwitkow_go_proto_validators.FieldError("RedeemDate", fmt.Errorf("message must exist"))
 	}
 	if this.RedeemDate != nil {
-		if err := go_proto_validators.CallValidatorIfExists(this.RedeemDate); err != nil {
-			return go_proto_validators.FieldError("RedeemDate", err)
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.RedeemDate); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("RedeemDate", err)
 		}
 	}
 	return nil
