@@ -3,8 +3,8 @@ package erc20_test
 import (
 	"testing"
 
-	examplecert "github.com/s7techlab/cckit/examples/cert"
 	"github.com/s7techlab/cckit/examples/erc20"
+	idtestdata "github.com/s7techlab/cckit/identity/testdata"
 	testcc "github.com/s7techlab/cckit/testing"
 	expectcc "github.com/s7techlab/cckit/testing/expect"
 
@@ -12,11 +12,19 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestCars(t *testing.T) {
+func TestErc20(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Cars Suite")
+	RunSpecs(t, "ERC-20 Suite")
 }
 
+var (
+	ids = idtestdata.MustIdentities(idtestdata.Certificates, idtestdata.DefaultMSP)
+
+	// load actor certificates
+	TokenOwner     = ids[0]
+	AccountHolder1 = ids[1]
+	Spender1       = ids[2]
+)
 var _ = Describe(`ERC-20`, func() {
 
 	const TokenSymbol = `HLF`
@@ -27,18 +35,9 @@ var _ = Describe(`ERC-20`, func() {
 	//Create chaincode mock
 	erc20fs := testcc.NewMockStub(`erc20`, erc20.NewErc20FixedSupply())
 
-	// load actor certificates
-	actors, err := testcc.IdentitiesFromFiles(`SOME_MSP`, map[string]string{
-		`token_owner`:     `s7techlab.pem`,
-		`account_holder1`: `victor-nosov.pem`,
-		//`accoubt_holder2`: `victor-nosov.pem`
-	}, examplecert.Content)
-	if err != nil {
-		panic(err)
-	}
 	BeforeSuite(func() {
 		// init token haincode
-		expectcc.ResponseOk(erc20fs.From(actors[`token_owner`]).Init(TokenSymbol, TokenName, TotalSupply, Decimals))
+		expectcc.ResponseOk(erc20fs.From(TokenOwner).Init(TokenSymbol, TokenName, TotalSupply, Decimals))
 	})
 
 	Describe("ERC-20 creation", func() {
@@ -58,13 +57,13 @@ var _ = Describe(`ERC-20`, func() {
 		It("Allow everyone to get owner's token balance", func() {
 			expectcc.PayloadInt(
 				erc20fs.Query(
-					`balanceOf`, actors[`token_owner`].MspId, actors[`token_owner`].GetID()), TotalSupply)
+					`balanceOf`, TokenOwner.MspID, TokenOwner.GetID()), TotalSupply)
 		})
 
 		It("Allow everyone to get holder's token balance", func() {
 			expectcc.PayloadInt(
 				erc20fs.Query(
-					`balanceOf`, actors[`account_holder1`].MspId, actors[`account_holder1`].GetID()), 0)
+					`balanceOf`, AccountHolder1.MspID, AccountHolder1.GetID()), 0)
 		})
 	})
 
@@ -72,31 +71,31 @@ var _ = Describe(`ERC-20`, func() {
 
 		It("Disallow to transfer token to same account", func() {
 			expectcc.ResponseError(
-				erc20fs.From(actors[`token_owner`]).Invoke(
-					`transfer`, actors[`token_owner`].MspId, actors[`token_owner`].GetID(), 100),
+				erc20fs.From(TokenOwner).Invoke(
+					`transfer`, TokenOwner.MspID, TokenOwner.GetID(), 100),
 				erc20.ErrForbiddenToTransferToSameAccount)
 		})
 
 		It("Disallow token holder with zero balance to transfer tokens", func() {
 			expectcc.ResponseError(
-				erc20fs.From(actors[`account_holder1`]).Invoke(
-					`transfer`, actors[`token_owner`].MspId, actors[`token_owner`].GetID(), 100),
+				erc20fs.From(AccountHolder1).Invoke(
+					`transfer`, TokenOwner.MspID, TokenOwner.GetID(), 100),
 				erc20.ErrNotEnoughFunds)
 		})
 
 		It("Allow token holder with non zero balance to transfer tokens", func() {
 			expectcc.PayloadInt(
-				erc20fs.From(actors[`token_owner`]).Invoke(
-					`transfer`, actors[`account_holder1`].MspId, actors[`account_holder1`].GetID(), 100),
+				erc20fs.From(TokenOwner).Invoke(
+					`transfer`, AccountHolder1.MspID, AccountHolder1.GetID(), 100),
 				TotalSupply-100)
 
 			expectcc.PayloadInt(
 				erc20fs.Query(
-					`balanceOf`, actors[`token_owner`].MspId, actors[`token_owner`].GetID()), TotalSupply-100)
+					`balanceOf`, TokenOwner.MspID, TokenOwner.GetID()), TotalSupply-100)
 
 			expectcc.PayloadInt(
 				erc20fs.Query(
-					`balanceOf`, actors[`account_holder1`].MspId, actors[`account_holder1`].GetID()), 100)
+					`balanceOf`, AccountHolder1.MspID, AccountHolder1.GetID()), 100)
 		})
 
 	})
@@ -107,24 +106,69 @@ var _ = Describe(`ERC-20`, func() {
 			expectcc.PayloadInt(
 				erc20fs.Query(
 					`allowance`,
-					actors[`token_owner`].MspId, actors[`token_owner`].GetID(),
-					actors[`account_holder1`].MspId, actors[`account_holder1`].GetID()), 0)
+					TokenOwner.MspID, TokenOwner.GetID(),
+					Spender1.MspID, Spender1.GetID()), 0)
 		})
 
 		It("Allow token owner to set transfer allowance", func() {
 			expectcc.ResponseOk(
-				erc20fs.From(actors[`token_owner`]).Invoke(
-					`approve`, actors[`account_holder1`].MspId, actors[`account_holder1`].GetID(), 10))
+				erc20fs.From(TokenOwner).Invoke(
+					`approve`, Spender1.MspID, Spender1.GetID(), 10))
 		})
 
 		It("Allow everyone to check token transfer allowance", func() {
 			expectcc.PayloadInt(
 				erc20fs.Query(
 					`allowance`,
-					actors[`token_owner`].MspId, actors[`token_owner`].GetID(),
-					actors[`account_holder1`].MspId, actors[`account_holder1`].GetID()), 10)
+					TokenOwner.MspID, TokenOwner.GetID(),
+					Spender1.MspID, Spender1.GetID()), 10)
 		})
 
-	})
+		It("Allow everyone to check token transfer allowance", func() {
+			expectcc.PayloadInt(
+				erc20fs.Query(
+					`allowance`,
+					TokenOwner.MspID, TokenOwner.GetID(),
+					Spender1.MspID, Spender1.GetID()), 10)
+		})
 
+		It("Disallow nonspender to initiate payment from owner waller", func() {
+			expectcc.ResponseError(
+				erc20fs.From(AccountHolder1).Invoke(
+					`transferFrom`,
+					TokenOwner.MspID, TokenOwner.GetID(),
+					AccountHolder1.MspID, AccountHolder1.GetID(), 10), erc20.ErrSpenderNotHaveAllowance)
+		})
+
+		It("Disallow spender to initiate payment more than allowance from wallet owner", func() {
+			expectcc.ResponseError(
+				erc20fs.From(AccountHolder1).Invoke(
+					`transferFrom`,
+					TokenOwner.MspID, TokenOwner.GetID(),
+					AccountHolder1.MspID, AccountHolder1.GetID(), 11), erc20.ErrSpenderNotHaveAllowance)
+		})
+
+		It("Allow spender to initiate payment from owner waller", func() {
+			expectcc.PayloadInt(
+				erc20fs.From(Spender1).Invoke(
+					`transferFrom`,
+					TokenOwner.MspID, TokenOwner.GetID(),
+					AccountHolder1.MspID, AccountHolder1.GetID(), 9), TotalSupply-100-9)
+
+			expectcc.PayloadInt(
+				erc20fs.Query(
+					`balanceOf`, AccountHolder1.MspID, AccountHolder1.GetID()), 100+9)
+
+			expectcc.PayloadInt(
+				erc20fs.Query(
+					`balanceOf`, TokenOwner.MspID, TokenOwner.GetID()), TotalSupply-100-9)
+
+			expectcc.PayloadInt(
+				erc20fs.Query(
+					`allowance`,
+					TokenOwner.MspID, TokenOwner.GetID(),
+					Spender1.MspID, Spender1.GetID()), 1)
+
+		})
+	})
 })
