@@ -67,7 +67,7 @@ func InvokeStateClean(c router.Context) (interface{}, error) {
 // InvokeValueByKeyPut router handler puts value in chaincode state with composite key,
 // created with key parts ([]string)
 func InvokeStatePut(c router.Context) (interface{}, error) {
-	key, err := state.StringKey(c.Stub(), c.Param(`key`).([]string))
+	key, err := state.KeyToString(c.Stub(), c.Param(`key`).([]string))
 	if err != nil {
 		return nil, errors.Wrap(err, `unable to create key`)
 	}
@@ -97,7 +97,7 @@ func QueryKeysList(c router.Context) (interface{}, error) {
 
 // QueryStateGet router handler returns state entry by key ([]string)
 func QueryStateGet(c router.Context) (interface{}, error) {
-	key, err := state.StringKey(c.Stub(), c.Param(`key`).([]string))
+	key, err := state.KeyToString(c.Stub(), c.Param(`key`).([]string))
 	if err != nil {
 		return nil, errors.Wrap(err, `unable to create key`)
 	}
@@ -106,7 +106,7 @@ func QueryStateGet(c router.Context) (interface{}, error) {
 
 // QueryStateGet router handler delete state entry by key ([]string)
 func InvokeStateDelete(c router.Context) (interface{}, error) {
-	key, err := state.StringKey(c.Stub(), c.Param(`key`).([]string))
+	key, err := state.KeyToString(c.Stub(), c.Param(`key`).([]string))
 	if err != nil {
 		return nil, errors.Wrap(err, `unable to create key`)
 	}
