@@ -1,20 +1,12 @@
 package encryption
 
 import (
-	"github.com/hyperledger/fabric/bccsp/factory"
-	"github.com/hyperledger/fabric/core/chaincode/shim/ext/entities"
 	"github.com/pkg/errors"
 	"github.com/s7techlab/cckit/convert"
 	"github.com/s7techlab/cckit/router"
 )
 
 const TransientMapKey = `ENCODE_KEY`
-
-func init() {
-	if err := factory.InitFactories(nil); err != nil {
-		panic(err)
-	}
-}
 
 // EncryptArgs convert args to [][]byte and encrypt args with key
 func EncryptArgs(key []byte, args ...interface{}) ([][]byte, error) {
@@ -30,7 +22,7 @@ func EncryptArgs(key []byte, args ...interface{}) ([][]byte, error) {
 func EncryptArgsBytes(key []byte, argsBytes [][]byte) ([][]byte, error) {
 	eargs := make([][]byte, len(argsBytes))
 	for i, bb := range argsBytes {
-		encrypted, err := Encrypt(key, bb)
+		encrypted, err := EncryptBytes(key, bb)
 		if err != nil {
 			return nil, errors.Wrap(err, `encryption error`)
 		}
@@ -51,7 +43,7 @@ func DecryptArgs(key []byte, args [][]byte) ([][]byte, error) {
 			continue
 		}
 
-		decrypted, err := Decrypt(key, a)
+		decrypted, err := DecryptBytes(key, a)
 		if err != nil {
 			return nil, errors.Wrap(err, `decryption error`)
 		}
@@ -63,28 +55,20 @@ func DecryptArgs(key []byte, args [][]byte) ([][]byte, error) {
 // Encrypt converts value to []byte  and encrypts its with key
 func Encrypt(key []byte, value interface{}) ([]byte, error) {
 	// TODO: customize  IV
-	encrypter, err := entities.NewAES256EncrypterEntity("ID", factory.GetDefault(), key, make([]byte, 16))
-	if err != nil {
-		return nil, err
-	}
-
 	bb, err := convert.ToBytes(value)
 	if err != nil {
 		return nil, errors.Wrap(err, `convert values to bytes`)
 	}
 
-	return encrypter.Encrypt(bb)
+	return EncryptBytes(key, bb)
 }
 
 // Decrypt decrypts value with key
 func Decrypt(key, value []byte) ([]byte, error) {
-	encrypter, err := entities.NewAES256EncrypterEntity("ID", factory.GetDefault(), key, nil)
-	if err != nil {
-		return nil, err
-	}
+
 	bb := make([]byte, len(value))
 	copy(bb, value)
-	return encrypter.Decrypt(bb)
+	return DecryptBytes(key, bb)
 }
 
 // TransientMapWithKey creates transient map with encrypting/decrypting key
