@@ -5,8 +5,8 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	examplecert "github.com/s7techlab/cckit/examples/cert"
-	"github.com/s7techlab/cckit/identity"
+
+	"github.com/s7techlab/cckit/identity/testdata"
 	"github.com/s7techlab/cckit/router"
 	testcc "github.com/s7techlab/cckit/testing"
 	expectcc "github.com/s7techlab/cckit/testing/expect"
@@ -16,6 +16,10 @@ func TestPinger(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Pinger suite")
 }
+
+var (
+	Someone = testdata.Certificates[1].MustIdentity(`SOME_MSP`)
+)
 
 func New() *router.Chaincode {
 	r := router.New(`pingable`).
@@ -28,23 +32,19 @@ var _ = Describe(`Pinger`, func() {
 
 	//Create chaincode mock
 	cc := testcc.NewMockStub(`cars`, New())
-	invokerIdentity, err := identity.FromFile(`SOME_MSP`, `s7techlab.pem`, examplecert.Content)
-	if err != nil {
-		panic(err)
-	}
 
 	Describe("Pinger", func() {
 
 		It("Allow anynone to invoke ping method", func() {
 			//invoke chaincode method from authority actor
-			pingInfo := expectcc.PayloadIs(cc.From(invokerIdentity).Invoke(FuncPing), &PingInfo{}).(PingInfo)
-			Expect(pingInfo.InvokerID).To(Equal(invokerIdentity.GetID()))
-			Expect(pingInfo.InvokerCert).To(Equal(invokerIdentity.GetPEM()))
+			pingInfo := expectcc.PayloadIs(cc.From(Someone).Invoke(FuncPing), &PingInfo{}).(PingInfo)
+			Expect(pingInfo.InvokerID).To(Equal(Someone.GetID()))
+			Expect(pingInfo.InvokerCert).To(Equal(Someone.GetPEM()))
 
 			//check that we have event
 			pingInfoEvent := expectcc.EventPayloadIs(cc.ChaincodeEvent, &PingInfo{}).(PingInfo)
-			Expect(pingInfoEvent.InvokerID).To(Equal(invokerIdentity.GetID()))
-			Expect(pingInfoEvent.InvokerCert).To(Equal(invokerIdentity.GetPEM()))
+			Expect(pingInfoEvent.InvokerID).To(Equal(Someone.GetID()))
+			Expect(pingInfoEvent.InvokerCert).To(Equal(Someone.GetPEM()))
 		})
 
 	})
