@@ -1,23 +1,15 @@
 // Code generated by protoc-gen-gogo. DO NOT EDIT.
 // source: chaincode.proto
 
-/*
-Package service is a generated protocol buffer package.
-
-It is generated from these files:
-	chaincode.proto
-
-It has these top-level messages:
-	ChaincodeInput
-	ChaincodeLocator
-*/
 package service
 
-import proto "github.com/golang/protobuf/proto"
-import fmt "fmt"
-import math "math"
-import _ "github.com/hyperledger/fabric/protos/peer"
-import _ "github.com/hyperledger/fabric/protos/peer"
+import (
+	fmt "fmt"
+	proto "github.com/golang/protobuf/proto"
+	_ "github.com/hyperledger/fabric/protos/peer"
+	github_com_mwitkow_go_proto_validators "github.com/mwitkow/go-proto-validators"
+	math "math"
+)
 
 // Reference imports to suppress errors if they are not otherwise used.
 var _ = proto.Marshal
@@ -29,5 +21,13 @@ func (this *ChaincodeInput) Validate() error {
 	return nil
 }
 func (this *ChaincodeLocator) Validate() error {
+	return nil
+}
+func (this *ChaincodeExec) Validate() error {
+	if this.Input != nil {
+		if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(this.Input); err != nil {
+			return github_com_mwitkow_go_proto_validators.FieldError("Input", err)
+		}
+	}
 	return nil
 }
