@@ -3,15 +3,13 @@ package router_test
 import (
 	"testing"
 
-	"github.com/hyperledger/fabric/core/chaincode/shim"
-
-	"github.com/hyperledger/fabric/protos/peer"
-
-	testcc "github.com/s7techlab/cckit/testing"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
+	"github.com/hyperledger/fabric/core/chaincode/shim"
+	"github.com/hyperledger/fabric/protos/peer"
 	"github.com/s7techlab/cckit/router"
+	testcc "github.com/s7techlab/cckit/testing"
 )
 
 func TestRouter(t *testing.T) {
