@@ -6,7 +6,7 @@ var (
 	// ErrEntryTypeNotSupported entry type has no appropriate mapper type
 	ErrEntryTypeNotSupported = errors.New(`entry type not supported for mapping`)
 
-	// ErrStateMappingNotFound  occurs when mapping for state entry is not defined
+	// ErrStateMappingNotFound occurs when mapping for state entry is not defined
 	ErrStateMappingNotFound = errors.New(`state mapping not found`)
 
 	// ErrEventMappingNotFound occurs when mapping for event is not defined
@@ -19,4 +19,10 @@ var (
 
 	ErrFieldNotExists         = errors.New(`field is not exists`)
 	ErrPrimaryKeyerNotDefined = errors.New(`primary keyer is not defined`)
+
+	// ErrIndexAlreadyExists occurs when when trying to add index to mapping with existent name
+	ErrIndexAlreadyExists = errors.New(`index already exists`)
+
+	// ErrIndexReferenceNotFound occurs when trying to find entry by index
+	ErrIndexReferenceNotFound = errors.New(`index reference not found`)
 )
