@@ -4,11 +4,13 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/golang/protobuf/ptypes"
-	"github.com/hyperledger/fabric/protos/peer"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	examplecert "github.com/s7techlab/cckit/examples/cert"
+
+	"github.com/golang/protobuf/proto"
+	"github.com/golang/protobuf/ptypes"
+	"github.com/hyperledger/fabric/protos/peer"
+	identitytestdata "github.com/s7techlab/cckit/identity/testdata"
 	"github.com/s7techlab/cckit/state"
 	"github.com/s7techlab/cckit/state/mapping"
 	"github.com/s7techlab/cckit/state/mapping/testdata"
@@ -24,167 +26,120 @@ func TestState(t *testing.T) {
 }
 
 var (
-	actors                          testcc.Identities
-	protoCC, complexIDCC, sliceIDCC *testcc.MockStub
-	err                             error
+	compositeIDCC, complexIDCC, sliceIDCC, indexesCC *testcc.MockStub
+
+	Owner = identitytestdata.Certificates[0].MustIdentity(`SOME_MSP`)
 )
 var _ = Describe(`Mapping`, func() {
 
 	BeforeSuite(func() {
-		actors, err = testcc.IdentitiesFromFiles(`SOME_MSP`, map[string]string{
-			`owner`: `s7techlab.pem`,
-		}, examplecert.Content)
 
-		Expect(err).To(BeNil())
+		compositeIDCC = testcc.NewMockStub(`proto`, testdata.NewCompositeIdCC())
+		compositeIDCC.From(Owner).Init()
 
-		protoCC = testcc.NewMockStub(`cpapers`, testdata.NewProtoCC())
-		protoCC.From(actors[`owner`]).Init()
+		complexIDCC = testcc.NewMockStub(`complex_id`, testdata.NewComplexIdCC())
+		complexIDCC.From(Owner).Init()
 
-		complexIDCC = testcc.NewMockStub(`complexid`, testdata.NewComplexIdCC())
-		complexIDCC.From(actors[`owner`]).Init()
+		sliceIDCC = testcc.NewMockStub(`slice_id`, testdata.NewSliceIdCC())
+		sliceIDCC.From(Owner).Init()
 
-		sliceIDCC = testcc.NewMockStub(`sliceid`, testdata.NewSliceIdCC())
-		sliceIDCC.From(actors[`owner`]).Init()
+		indexesCC = testcc.NewMockStub(`indexes`, testdata.NewIndexesCC())
+		indexesCC.From(Owner).Init()
 	})
 
-	Describe(`Commercial paper extended, protobuf based schema with additional keys`, func() {
-		issueMock1 := testdata.ProtoIssueMocks[0]
-		issueMock2 := testdata.ProtoIssueMocks[1]
-		issueMock3 := testdata.ProtoIssueMocks[2]
-		issueMockExistingExternal := testdata.ProtoIssueMockExistingExternal
-		issueMockExistingPrimary := testdata.ProtoIssueMockExistingPrimary
+	Describe(`Entity with composite id`, func() {
+		create1 := testdata.CreateEntityWithCompositeId[0]
+		create2 := testdata.CreateEntityWithCompositeId[1]
+		create3 := testdata.CreateEntityWithCompositeId[2]
 
 		It("Allow to get mapping data by namespace", func() {
-			mapping, err := testdata.ProtoStateMapping.GetByNamespace(state.Key{`ProtoEntity`})
+			mapping, err := testdata.EntityWithCompositeIdStateMapping.GetByNamespace(state.Key{`EntityWithCompositeId`})
 			Expect(err).NotTo(HaveOccurred())
-			Expect(mapping.Schema()).To(BeEquivalentTo(&schema.ProtoEntity{}))
+			Expect(mapping.Schema()).To(BeEquivalentTo(&schema.EntityWithCompositeId{}))
 		})
 
 		It("Allow to add data to chaincode state", func(done Done) {
-			events := protoCC.EventSubscription()
-			expectcc.ResponseOk(protoCC.Invoke(`issue`, &issueMock1))
+			events := compositeIDCC.EventSubscription()
+			expectcc.ResponseOk(compositeIDCC.Invoke(`create`, create1))
 
 			Expect(<-events).To(BeEquivalentTo(&peer.ChaincodeEvent{
-				EventName: `IssueProtoEntity`,
-				Payload:   testcc.MustProtoMarshal(&issueMock1),
+				EventName: `CreateEntityWithCompositeId`,
+				Payload:   testcc.MustProtoMarshal(create1),
 			}))
 
-			expectcc.ResponseOk(protoCC.Invoke(`issue`, &issueMock2))
-			expectcc.ResponseOk(protoCC.Invoke(`issue`, &issueMock3))
+			expectcc.ResponseOk(compositeIDCC.Invoke(`create`, create2))
+			expectcc.ResponseOk(compositeIDCC.Invoke(`create`, create3))
 
 			close(done)
-		}, 0.2)
-
-		It("Disallow to insert entries with same uniq AND primary keys", func() {
-			expectcc.ResponseError(protoCC.Invoke(`issue`, &issueMock1))
 		})
 
-		It("Disallow to add data to chaincode state with same uniq key fields", func() {
-			// errored on checking uniq key
-			expectcc.ResponseError(
-				protoCC.Invoke(`issue`, &issueMockExistingExternal),
-				mapping.ErrMappingUniqKeyExists)
-		})
-
-		It("Disallow adding data to chaincode state with same primary key fields", func() {
-			// errored obn checkong primary key
-			expectcc.ResponseError(
-				protoCC.Invoke(`issue`, &issueMockExistingPrimary),
-				state.ErrKeyAlreadyExists)
+		It("Disallow to insert entries with same primary key", func() {
+			expectcc.ResponseError(compositeIDCC.Invoke(`create`, create1), state.ErrKeyAlreadyExists)
 		})
 
 		It("Allow to get entry list", func() {
-			entities := expectcc.PayloadIs(protoCC.Query(`list`),
-				&schema.ProtoEntityList{}).(*schema.ProtoEntityList)
+			entities := expectcc.PayloadIs(compositeIDCC.Query(`list`),
+				&schema.EntityWithCompositeIdList{}).(*schema.EntityWithCompositeIdList)
 			Expect(len(entities.Items)).To(Equal(3))
-			Expect(entities.Items[0].Name).To(Equal(issueMock1.Name))
-			Expect(entities.Items[0].Value).To(BeNumerically("==", 0))
-			Expect(entities.Items[0].ExternalId).To(Equal(issueMock1.ExternalId))
-		})
-
-		It("Allow finding data by uniq key", func() {
-
-			cpaperFromCCByExtID := expectcc.PayloadIs(
-				protoCC.Query(`getByExternalId`, issueMock1.ExternalId),
-				&schema.ProtoEntity{}).(*schema.ProtoEntity)
-
-			cpaperFromCC := expectcc.PayloadIs(
-				protoCC.Query(`get`, &schema.ProtoEntityId{
-					IdFirstPart:  issueMock1.IdFirstPart,
-					IdSecondPart: issueMock1.IdSecondPart},
-				),
-				&schema.ProtoEntity{}).(*schema.ProtoEntity)
-
-			Expect(cpaperFromCCByExtID).To(BeEquivalentTo(cpaperFromCC))
-		})
-
-		It("Allow to get idx state key by uniq key", func() {
-			idxKey, err := testdata.ProtoStateMapping.IdxKey(&schema.ProtoEntity{}, `ExternalId`, []string{issueMock1.ExternalId})
-			Expect(err).NotTo(HaveOccurred())
-
-			Expect(idxKey).To(BeEquivalentTo([]string{
-				mapping.KeyRefNamespace,
-				strings.Join(mapping.SchemaNamespace(&schema.ProtoEntity{}), `-`),
-				`ExternalId`,
-				issueMock1.ExternalId,
-			}))
-		})
-
-		It("Disallow finding data by non existent uniq key", func() {
-			expectcc.ResponseError(
-				protoCC.Query(`getByExternalId`, `some-non-existent-id`), `uniq index`)
+			Expect(entities.Items[0].Name).To(Equal(create1.Name))
+			Expect(entities.Items[0].Value).To(BeNumerically("==", create1.Value))
 		})
 
 		It("Allow to get entry raw protobuf", func() {
-			cpaperProtoFromCC := protoCC.Query(`get`,
-				&schema.ProtoEntityId{
-					IdFirstPart:  issueMock1.IdFirstPart,
-					IdSecondPart: issueMock1.IdSecondPart},
+			dataFromCC := compositeIDCC.Query(`get`,
+				&schema.EntityCompositeId{
+					IdFirstPart:  create1.IdFirstPart,
+					IdSecondPart: create1.IdSecondPart},
 			).Payload
 
-			stateProtoEntity := &schema.ProtoEntity{
-				IdFirstPart:  issueMock1.IdFirstPart,
-				IdSecondPart: issueMock1.IdSecondPart,
-				Name:         issueMock1.Name,
-				Value:        0,
-				ExternalId:   issueMock1.ExternalId,
+			e := &schema.EntityWithCompositeId{
+				IdFirstPart:  create1.IdFirstPart,
+				IdSecondPart: create1.IdSecondPart,
+				Name:         create1.Name,
+				Value:        create1.Value,
 			}
-			Expect(cpaperProtoFromCC).To(Equal(testcc.MustProtoMarshal(stateProtoEntity)))
+			Expect(dataFromCC).To(Equal(testcc.MustProtoMarshal(e)))
 		})
 
 		It("Allow update data in chaincode state", func() {
-			expectcc.ResponseOk(protoCC.Invoke(`increment`, &schema.IncrementProtoEntity{
-				IdFirstPart:  issueMock1.IdFirstPart,
-				IdSecondPart: issueMock1.IdSecondPart,
-				Name:         issueMock1.Name,
+			expectcc.ResponseOk(compositeIDCC.Invoke(`update`, &schema.UpdateEntityWithCompositeId{
+				IdFirstPart:  create1.IdFirstPart,
+				IdSecondPart: create1.IdSecondPart,
+				Name:         `New name`,
+				Value:        1000,
 			}))
 
 			entityFromCC := expectcc.PayloadIs(
-				protoCC.Query(`get`, &schema.ProtoEntityId{
-					IdFirstPart:  issueMock1.IdFirstPart,
-					IdSecondPart: issueMock1.IdSecondPart,
+				compositeIDCC.Query(`get`, &schema.EntityCompositeId{
+					IdFirstPart:  create1.IdFirstPart,
+					IdSecondPart: create1.IdSecondPart,
 				}),
-				&schema.ProtoEntity{}).(*schema.ProtoEntity)
+				&schema.EntityWithCompositeId{}).(*schema.EntityWithCompositeId)
 
 			// state is updated
-			Expect(entityFromCC.Value).To(BeNumerically("==", 1))
+			Expect(entityFromCC.Name).To(Equal(`New name`))
+			Expect(entityFromCC.Value).To(BeNumerically("==", 1000))
 		})
 
 		It("Allow to delete entry", func() {
-			toDelete := &schema.ProtoEntityId{
-				IdFirstPart:  issueMock1.IdFirstPart,
-				IdSecondPart: issueMock1.IdSecondPart,
+			toDelete := &schema.EntityCompositeId{
+				IdFirstPart:  create1.IdFirstPart,
+				IdSecondPart: create1.IdSecondPart,
 			}
 
-			expectcc.ResponseOk(protoCC.Invoke(`delete`, toDelete))
-			cpapers := expectcc.PayloadIs(
-				protoCC.Invoke(`list`),
-				&schema.ProtoEntityList{},
-			).(*schema.ProtoEntityList)
+			expectcc.ResponseOk(compositeIDCC.Invoke(`delete`, toDelete))
+			ee := expectcc.PayloadIs(
+				compositeIDCC.Invoke(`list`),
+				&schema.EntityWithCompositeIdList{}).(*schema.EntityWithCompositeIdList)
 
-			Expect(len(cpapers.Items)).To(Equal(2))
-			expectcc.ResponseError(protoCC.Invoke(`get`, toDelete), state.ErrKeyNotFound)
+			Expect(len(ee.Items)).To(Equal(2))
+			expectcc.ResponseError(compositeIDCC.Invoke(`get`, toDelete), state.ErrKeyNotFound)
 		})
+
+		It("Allow to insert entry once more time", func() {
+			expectcc.ResponseOk(compositeIDCC.Invoke(`create`, create1))
+		})
+
 	})
 
 	Describe(`Entity with complex id`, func() {
@@ -193,7 +148,7 @@ var _ = Describe(`Mapping`, func() {
 
 		It("Allow to add data to chaincode state", func() {
 			expectcc.ResponseOk(complexIDCC.Invoke(`entityInsert`, ent1))
-			keys := expectcc.PayloadIs(complexIDCC.From(actors[`owner`]).Invoke(
+			keys := expectcc.PayloadIs(complexIDCC.From(Owner).Invoke(
 				`debugStateKeys`, []string{`EntityWithComplexId`}), &[]string{}).([]string)
 			Expect(len(keys)).To(Equal(1))
 
@@ -223,7 +178,7 @@ var _ = Describe(`Mapping`, func() {
 
 		It("Allow to add data to chaincode state", func() {
 			expectcc.ResponseOk(sliceIDCC.Invoke(`entityInsert`, ent2))
-			keys := expectcc.PayloadIs(sliceIDCC.From(actors[`owner`]).Invoke(
+			keys := expectcc.PayloadIs(sliceIDCC.From(Owner).Invoke(
 				`debugStateKeys`, []string{`EntityWithSliceId`}), &[]string{}).([]string)
 
 			Expect(len(keys)).To(Equal(1))
@@ -245,5 +200,139 @@ var _ = Describe(`Mapping`, func() {
 
 			Expect(listFromCC.Items[0].Value).To(Equal(testcc.MustProtoMarshal(ent2)))
 		})
+	})
+
+	Describe(`Entity with indexes`, func() {
+
+		create1 := testdata.CreateEntityWithIndexes[0]
+		create2 := testdata.CreateEntityWithIndexes[1]
+
+		It("Allow to add data with single external id", func() {
+			expectcc.ResponseOk(indexesCC.Invoke(`create`, create1))
+		})
+
+		It("Disallow to add data to chaincode state with same uniq key fields", func() {
+			createWithNewId := proto.Clone(create1).(*schema.CreateEntityWithIndexes)
+			createWithNewId.Id = `abcdef` // id is really new
+
+			// errored on checking uniq key
+			expectcc.ResponseError(
+				indexesCC.Invoke(`create`, create1),
+				mapping.ErrMappingUniqKeyExists)
+		})
+
+		It("Allow finding data by uniq key", func() {
+			fromCCByExtId := expectcc.PayloadIs(
+				indexesCC.Query(`getByExternalId`, create1.ExternalId),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			fromCCById := expectcc.PayloadIs(
+				indexesCC.Query(`get`, create1.Id),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			Expect(fromCCByExtId).To(BeEquivalentTo(fromCCById))
+		})
+
+		It("Allow to get idx state key by uniq key", func() {
+			idxKey, err := testdata.EntityWithIndexesStateMapping.IdxKey(
+				&schema.EntityWithIndexes{}, `ExternalId`, []string{create1.ExternalId})
+			Expect(err).NotTo(HaveOccurred())
+
+			Expect(idxKey).To(BeEquivalentTo([]string{
+				mapping.KeyRefNamespace,
+				strings.Join(mapping.SchemaNamespace(&schema.EntityWithIndexes{}), `-`),
+				`ExternalId`,
+				create1.ExternalId,
+			}))
+		})
+
+		It("Disallow finding data by non existent uniq key", func() {
+			expectcc.ResponseError(
+				indexesCC.Query(`getByExternalId`, `some-non-existent-id`),
+				mapping.ErrIndexReferenceNotFound)
+		})
+
+		It("Allow to add data with multiple external id", func() {
+			expectcc.ResponseOk(indexesCC.Invoke(`create`, create2))
+		})
+
+		It("Allow to find data by multi key", func() {
+			fromCCByExtId1 := expectcc.PayloadIs(
+				indexesCC.Query(`getByOptMultiExternalId`, create2.OptionalExternalIds[0]),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			fromCCByExtId2 := expectcc.PayloadIs(
+				indexesCC.Query(`getByOptMultiExternalId`, create2.OptionalExternalIds[1]),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			fromCCById := expectcc.PayloadIs(
+				indexesCC.Query(`get`, create2.Id),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			Expect(fromCCByExtId1).To(BeEquivalentTo(fromCCById))
+			Expect(fromCCByExtId2).To(BeEquivalentTo(fromCCById))
+		})
+
+		It("Allow update indexes value", func() {
+			update2 := &schema.UpdateEntityWithIndexes{
+				Id:                  create2.Id,
+				ExternalId:          `some_new_external_id`,
+				OptionalExternalIds: []string{create2.OptionalExternalIds[0], `AND SOME NEW`},
+			}
+			expectcc.ResponseOk(indexesCC.Invoke(`update`, update2))
+		})
+
+		It("Allow to find data by updated multi key", func() {
+			fromCCByExtId1 := expectcc.PayloadIs(
+				indexesCC.Query(`getByOptMultiExternalId`, create2.OptionalExternalIds[0]),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			fromCCByExtId2 := expectcc.PayloadIs(
+				indexesCC.Query(`getByOptMultiExternalId`, `AND SOME NEW`),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			Expect(fromCCByExtId1.Id).To(Equal(create2.Id))
+			Expect(fromCCByExtId2.Id).To(Equal(create2.Id))
+
+			Expect(fromCCByExtId2.OptionalExternalIds).To(
+				BeEquivalentTo([]string{create2.OptionalExternalIds[0], `AND SOME NEW`}))
+		})
+
+		It("Disallow to find data by previous multi key", func() {
+			expectcc.ResponseError(
+				indexesCC.Query(`getByOptMultiExternalId`, create2.OptionalExternalIds[1]),
+				mapping.ErrIndexReferenceNotFound)
+		})
+
+		It("Allow to find data by updated uniq key", func() {
+			fromCCByExtId := expectcc.PayloadIs(
+				indexesCC.Query(`getByExternalId`, `some_new_external_id`),
+				&schema.EntityWithIndexes{}).(*schema.EntityWithIndexes)
+
+			Expect(fromCCByExtId.Id).To(Equal(create2.Id))
+			Expect(fromCCByExtId.ExternalId).To(Equal(`some_new_external_id`))
+		})
+
+		It("Disallow to find data by previous uniq key", func() {
+			expectcc.ResponseError(
+				indexesCC.Query(`getByExternalId`, create2.ExternalId),
+				mapping.ErrIndexReferenceNotFound)
+		})
+
+		It("Allow to delete entry", func() {
+			expectcc.ResponseOk(indexesCC.Invoke(`delete`, create2.Id))
+
+			ee := expectcc.PayloadIs(
+				indexesCC.Invoke(`list`),
+				&schema.EntityWithIndexesList{}).(*schema.EntityWithIndexesList)
+
+			Expect(len(ee.Items)).To(Equal(1))
+			expectcc.ResponseError(indexesCC.Invoke(`get`, create2.Id), state.ErrKeyNotFound)
+		})
+
+		It("Allow to insert entry once more time", func() {
+			expectcc.ResponseOk(indexesCC.Invoke(`create`, create2))
+		})
+
 	})
 })
